% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_solve.R
\name{add_residual}
\alias{add_residual}
\title{Convolute the First Passage Times with the Non-Decision Time Distribution}
\usage{
add_residual(pdf_nt, pdf_u, pdf_l, dt, nt)
}
\arguments{
\item{pdf_nt}{the non-decision time density values}

\item{pdf_u, pdf_l}{the first passage times}

\item{dt, nt}{step size and number of steps for the time space (for input
checks and scaling)}
}
\value{
a list of PDFs for one condition "pdf_u" and "pdf_l"
}
\description{
Calls \link[stats:convolve]{stats::convolve} for the first passage times and the non-decision
time distribution to derive the full distribution of response times. After
convolution, I add the robustness parameter.
}
\keyword{internal}
