% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_stats.R
\name{aggregate_stats}
\alias{aggregate_stats}
\title{Aggregate Statistics Across ID}
\usage{
aggregate_stats(stat_df)
}
\arguments{
\item{stat_df}{A \code{data.frame} of class \code{stats_dm} (see
\code{\link[=new_stats_dm]{new_stats_dm()}})}
}
\value{
Returns the statistics aggregated across the relevant cols.
}
\description{
\code{aggregate_stats} is a (not exported) function to aggregate
\code{stats_dm} objects across \code{ID}s. Since the column names may vary by the
statistic type, the behavior of aggregate depends on the subclass of
\code{stats_dm}.
}
\details{
\code{aggregate_stats} calls the \code{\link[=internal_aggregate]{internal_aggregate()}} with the
relevant arguments
}
\seealso{
\link{new_stats_dm}, \link{calc_stats},
\code{\link[=internal_aggregate]{internal_aggregate()}}
}
\keyword{internal}
