% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{b_hyperbol}
\alias{b_hyperbol}
\alias{dt_b_hyperbol}
\title{Collapsing Boundary - Hyperbolic Ratio Function}
\usage{
b_hyperbol(prms_model, prms_solve, t_vec, one_cond, ddm_opts)

dt_b_hyperbol(prms_model, prms_solve, t_vec, one_cond, ddm_opts)
}
\arguments{
\item{prms_model}{the model parameters, containing b0, kappa, t05}

\item{prms_solve}{solver settings}

\item{t_vec}{time space}

\item{one_cond}{one condition}

\item{ddm_opts}{optional arguments attached to an object}
}
\value{
a vector of the same length as t_vec with the boundary values (or
the deriviative) for each element of the vector.
}
\description{
Provides the boundary, collapsing as a hyperbolic ratio function.
}
\details{
\code{b_hyperbol} and \code{dt_b_hyperbol} provide the plain boundary values and the
respective derivative, respectively.
}
\keyword{internal}
