% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{b_weibull}
\alias{b_weibull}
\alias{dt_b_weibull}
\title{Collapsing Boundary - Weibull Function}
\usage{
b_weibull(prms_model, prms_solve, t_vec, one_cond, ddm_opts)

dt_b_weibull(prms_model, prms_solve, t_vec, one_cond, ddm_opts)
}
\arguments{
\item{prms_model}{the model parameters, containing b0, lambda, k, kappa}

\item{prms_solve}{solver settings}

\item{t_vec}{time space}

\item{one_cond}{one condition}

\item{ddm_opts}{optional arguments attached to an object}
}
\value{
a vector of the same length as t_vec with the boundary values (or
the deriviative) for each element of the vector.
}
\description{
Provides the boundary, collapsing in accordance with a Weibull function
}
\details{
\code{b_weibull} and \code{dt_b_weibull} provide the plain boundary values and the
respective derivative, respectively.
}
