% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_stats.R
\name{calc_cafs}
\alias{calc_cafs}
\title{Calcuating CAFs}
\usage{
calc_cafs(drift_dm_obj, type = "obs", n_bins = 5)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}}

\item{type}{character, indicating whether CAFs of the observed data ("obs"),
of the model's predictions ("pred"), or both ("both") should be calculated.
Default is "obs".}

\item{n_bins}{numeric, providing the number of bins that should be used when
calculating the CAFs}
}
\value{
Returns a data.frame containing at least the columns \code{Cond}, \code{Bin}, and
\code{P_Corr} where \code{P_Corr} is the "percentage of correct responses". In case
\code{type = "both"}, then the data.frame contains an additional column \code{Source}
identifying whether an entry describes the observed data ("obs") or the
model's predictions ("pred").
}
\description{
This function calculates the conditional accuracy functions (CAFs) of the
model and/or the observed data.
}
\details{
CAFs are a way to quantify response accuracy against speed. To calculate
CAFs, RTs (whether correct or incorrect) are first binned and then the
percentage of correct responses per bin is calculated.

When calculating model-based CAFs, a joint cdf combining both the pdf
of correct and incorrect responses is calculated. Afterwards, this cdf
is separated into even-spaced segments and the contribution of
the pdf associated with a correct response relative to the joint cdf is
calculated
}
