% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_stats.R
\name{calc_quantiles}
\alias{calc_quantiles}
\title{Calcuating Quantiles}
\usage{
calc_quantiles(drift_dm_obj, type = "obs", probs = seq(0.1, 0.9, 0.1))
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}}

\item{type}{character, indicating whether quantiles of the observed data
("obs"), of the model's predicted pdfs ("pred"), or both ("both") should
be calculated. Default is "obs".}

\item{probs}{numeric vector, providing the probabilites with values in
\eqn{\[0, 1s\]} for which quantiles should be calculated. Default is
\code{seq(0.1, 0.9, 0.1)}.}
}
\value{
Returns a data.frame containing at least the columns \code{Cond}, \code{Prob},
\code{Quant_Corr} and \code{Quant_Err}. Here, \code{Cond} indicate the conditions of the
model, \code{Prob} the probabilites, and \code{Quant_Corr} and \code{Quant_Err} the
quantiles associated with correct and error responses, respectively.
In case \code{type = "both"}, then the returned data.frame contains an additional
column \code{Source} identifying whether an entry describes the observed data
("obs") or the model's predictions ("pred").
}
\description{
This function calculates the quantiles of the
model's predicted pdfs and/or of the observed response times.
}
\details{
For observed response times, the function \link[stats:quantile]{stats::quantile} is used with
default settings
}
