% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_cost_functions.R
\name{calc_rmse}
\alias{calc_rmse}
\title{Compute RMSE from predicted and observed quantiles/CAFs}
\usage{
calc_rmse(quants_pred, cafs_pred, quants_obs, cafs_obs, weight_err = 1.5)
}
\description{
Compute RMSE from predicted and observed quantiles/CAFs
}
\keyword{internal}
