% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_stats.R
\name{calc_stats}
\alias{calc_stats}
\title{Calcuating Statistics}
\usage{
calc_stats(drift_dm_obj, type, source = "both", ...)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}}

\item{type}{character vector, indicating which statistics should be
calculated. Currently supported options are "cafs" or "quantiles".}

\item{source}{character, indicating whether CAFs of the observed data ("obs"),
of the model's predictions ("pred"), or both ("both") should be calculated.
Default is "both".}

\item{...}{additional parameters passed down to the functions handling the
calculations. See Details for more information}

\item{n_bins}{numeric, providing the number of bins that should be used when
calculating the CAFs}
}
\value{
The return value depends on the \code{type} argument. If \code{type} is  a charcter
vector of length 1, a single data.frame is returned containing the requested
statistics. If \code{type} is a character vector of length > 1, a named list
is passed back, with each entry in the list corresponding to a data.frame of
statistics.
}
\description{
This function is a wrapper for calculating statistics on model predictions
or observed data. Currently supported are:
\itemize{
\item Conditional Accuracy Functions (CAFs)
\item Quantiles
}
}
\section{Conditional Accuracy Function (CAFs)}{
CAFs are a way to quantify response accuracy against speed. To calculate
CAFs, RTs (whether correct or incorrect) are first binned and then the
percentage of correct responses per bin is calculated.

When calculating model-based CAFs, a joint cdf combining both the pdf
of correct and incorrect responses is calculated. Afterwards, this cdf
is separated into even-spaced segments and the contribution of
the pdf associated with a correct response relative to the joint cdf is
calculated.

The number of bins can be controlled by passing the argument \code{n_bins}.
The default is 5.
}

\section{Quantiles}{
For observed response times, the function \link[stats:quantile]{stats::quantile} is used with
default settings.

Which quantiles are calcuated can be controlled by providing the
probabilites with values in \eqn{\[0, 1s\]}. Default is
\code{seq(0.1, 0.9, 0.1)}.
}

