% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dm.R
\name{check_b_coding}
\alias{check_b_coding}
\title{Check a B_Coding}
\usage{
check_b_coding(b_coding)
}
\arguments{
\item{b_coding}{a named list}
}
\value{
the unmodified list for convenience
}
\description{
Checks if a list satisfies the requirements to be considered a valid
\link{b_coding}
}
\details{
Checks for...
\itemize{
\item input being a list
\item list is of length three and provides the names "column", "u_name_value",
"l_name_value",
\item if b_coding$column provides a single string
\item if b_coding$u_name_value and b_coding$l_name_value provides a single
named value of type character or numeric, and that both are of the same
type
}
}
