% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_flex_prms.R
\name{check_digit_larger_0}
\alias{check_digit_larger_0}
\title{Checks if a variable/vector of length 1 is a number > 0 or 0/expression}
\usage{
check_digit_larger_0(x)
}
\arguments{
\item{x}{a vector of length 1}
}
\value{
a single logical
}
\description{
assumed that x is either a digit or an expression
}
\keyword{internal}
