% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_flex_prms.R
\name{check_internal_list}
\alias{check_internal_list}
\title{checks if all entries of internal_list are an expression or integer.
Throws an error if not (internal docu)}
\usage{
check_internal_list(internal_list)
}
\arguments{
\item{internal_list}{}
}
\description{
checks if all entries of internal_list are an expression or integer.
Throws an error if not (internal docu)
}
