% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{component_shelf}
\alias{component_shelf}
\title{Diffusion model components}
\usage{
component_shelf()
}
\description{
The following function is meant as a convenient way to access pre-built
model component functions. It returns a list of the following functions.

\code{mu_constant}, this function provides the component function for a constant
drift rate with parameter \code{muc}.

\code{mu_dmc}, this function provides the superimposed diffusion process
of DMC. Necessary parameters are \code{muc} (drift rate of the controlled
process), \code{a} (shape..), \code{A} (amplitude...), \code{tau} (and scale of the
automatic process.

\code{mu_int_constant}, provides the complementary integral to \code{mu_constant}.

\code{mu_int_dmc}, provides the complementary integral to \code{mu_dmc}.

\code{x_dirac_0}, this function provides a dirac delta for a starting point
centered between the boundaries

\code{x_uniform}, this function provides a uniform distribution for a start point
centered between the boundaries. Requires a parameter \code{range_start}

\code{x_beta}, this function provides the function component for a symmetric
beta-shaped starting point distribution with parameter \code{alpha}.

\code{b_constant}, this function provides a constant
boundary with parameter \code{b}.

\code{nt_constant}, this function provides a constant
non-decision time with parameter \code{non_dec}.

\code{nt_uniform}, this function provides a uniform distribution for the
non-decision time. Requires the parameters \code{non_dec} and \code{range_non_dec}.

\code{nt_truncated_normal}, this function provides the component function for
a normally distributed non-decision time with parameters \code{non_dec},
\code{sd_non_dec}. The Distribution is truncated to \eqn{[0, t_max]}.
}
\details{
See \code{vignette("use_ddm_models", "dRiftDM")} for more information on how
to set/modify/customize the components of a diffusion model.
}
