% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_bayes_estimate.R
\name{create_temperatures}
\alias{create_temperatures}
\title{Create "Temperatures" for TIDE}
\usage{
create_temperatures(n_chains, algorithm)
}
\arguments{
\item{n_chains}{numeric}

\item{algorithm}{"tide" or anything else}
}
\value{
a numeric vector of length equal to \code{n_chains}. The returned values
correspond to quantiles of a Beta(0.3, 1) distribution for
\code{algorithm == "tide"}. Otherwise a numeric vector of \code{1}s is returned.
}
\description{
Create "Temperatures" for TIDE
}
\keyword{internal}
