% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_bayes_estimate.R
\name{crossover}
\alias{crossover}
\title{Perform Crossover Between Chains}
\usage{
crossover(which, ...)
}
\arguments{
\item{which}{character string, Either \code{"diff"} or \code{"migration"}.}

\item{...}{Further arguments passed to the underlying crossover function.}
}
\value{
A list with the following components:
\itemize{
\item \code{new_prms_across_chains}: The updated parameter matrix of shape \verb{p × n}.
\item \code{new_pis_across_chains}: The updated vector of log-posterior values.
\item \code{new_log_likes_across_chains}: The updated vector of log-likelihood values.
}
}
\description{
This function dispatches to either \code{\link[=full_crossover]{full_crossover()}} or
\code{\link[=migration_crossover]{migration_crossover()}} depending on the \code{which} argument.
}
\keyword{internal}
