% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_stats.R
\name{do_resampling}
\alias{do_resampling}
\title{Temporarily suppress new stats generation during resampling}
\usage{
do_resampling(x)
}
\arguments{
\item{x}{An expression to evaluate.}
}
\value{
The result of evaluated \code{x}.
}
\description{
Internal helper that sets the \code{skip_new_stats_dm} and
\code{skip_validate_stats_dm} options to \code{TRUE} before
evaluating an expression, and resets them to \code{NULL} afterward (see also
\code{\link[=stats.options]{stats.options()}}). Intended to
prevent the (unncessary) creation/checking of stats_dm objects during
resampling.
}
\keyword{internal}
