% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drift_dm_cost_functions}
\alias{drift_dm_cost_functions}
\title{Available Cost Functions for Model Estimation}
\usage{
drift_dm_cost_functions()
}
\value{
A character vector of cost function names.
}
\description{
Returns the names of implemented cost functions.
}
\keyword{internal}
