% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drift_dm_pre_built_models}
\alias{drift_dm_pre_built_models}
\title{Pre-built Drift Diffusion Models}
\usage{
drift_dm_pre_built_models()
}
\value{
A character vector of model names.
}
\description{
Returns the names of available pre-built DDMs in dRiftDM.
}
\keyword{internal}
