% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drift_dm_stats_types}
\alias{drift_dm_stats_types}
\title{Available types of statistics}
\usage{
drift_dm_stats_types(context = NULL)
}
\arguments{
\item{context}{a character string, indicating the context. If
\code{NULL}, all available types are returned.}
}
\value{
a character vector of valid statistic types for the given context.
}
\description{
Internal helper to return supported statistic types depending on the
context (e.g., for observed data.frames or fitted model objects).
}
\keyword{internal}
