% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dm.R
\name{drop_levels_ID_column}
\alias{drop_levels_ID_column}
\title{Maybe droplevels of ID column}
\usage{
drop_levels_ID_column(some_data)
}
\arguments{
\item{some_data}{a data.frame with an ID column}
}
\value{
if the ID column is not of type factor, then the unmodified object is
returned.

if the ID column is of type factor, \link{droplevels} is applied, and if levels
were dropped, a warning is ushered
}
\description{
This function takes a data frame with an ID colmumn, and drops the unused
levels from the ID column if it is factor; in this case a warning is
ushered
}
\keyword{internal}
