% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{dt_b_constant}
\alias{dt_b_constant}
\title{Derivative of a Constant Boundary}
\usage{
dt_b_constant(prms_model, prms_solve, t_vec, one_cond, ddm_opts)
}
\arguments{
\item{prms_model}{the model parameters, no prm label required}

\item{prms_solve}{solver settings}

\item{t_vec}{time space}

\item{one_cond}{one condition}

\item{ddm_opts}{optional arguments attached to an object}
}
\value{
a vector of the same length as t_vec only zeros.
}
\description{
Derivative of a Constant Boundary
}
\keyword{internal}
