% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_estimate.R
\name{estimate_model}
\alias{estimate_model}
\title{Estimate the parameters of a drift_dm model}
\usage{
estimate_model(
  drift_dm_obj,
  lower,
  upper,
  verbose = FALSE,
  use_de_optim = TRUE,
  use_nmkb = FALSE,
  seed = NULL,
  de_n_cores = 1,
  de_control = list(reltol = 1e-09, steptol = 50, itermax = 200),
  nmkb_control = list(tol = 1e-06)
)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}}

\item{lower, upper}{numeric vectors specifying the lower and upper bounds on
each parameter to be optimized. Their length have to match the number of
\code{free_prms} in the model. Note that the i-th \code{lower} and \code{upper} element
refers to the i-th parameter listed in \code{free_prms} of \code{drift_dm_obj}.}

\item{verbose}{logical, indicating if the parameter values and the negative
log-likelihood should be printed at each evaluation of the model. Default is
\code{FALSE}.}

\item{use_de_optim}{logical, indicating whether Differential Evolution via
\link[DEoptim:DEoptim]{DEoptim::DEoptim} should be used. Default is \code{TRUE}}

\item{use_nmkb}{logical, indicating whether Nelder-Mead via
\link[dfoptim:nmkb]{dfoptim::nmkb} should be used. Default is \code{FALSE}. If both \code{use_de_optim}
and \code{use_nmkb} are \code{TRUE}, then Nelder-Mead follows Differential Evolution}

\item{seed}{a single numeric, providing a seed for the Differential Evolution
algorithm}

\item{de_n_cores}{a single numeric, indicating the number of cores to use.
Run \code{\link[parallel:detectCores]{parallel::detectCores()}} to see how many cores are available on your
machine. Note that it is generally not recommended to use all of your cores
as this will drastically slow down your machine for any additional task.}

\item{de_control, nmkb_control}{lists of additional control parameters passed
to \link[DEoptim:DEoptim]{DEoptim::DEoptim} and \link[dfoptim:nmkb]{dfoptim::nmkb}. Default settings will lead
\link[DEoptim:DEoptim]{DEoptim::DEoptim} to stop if the algorithm is unable to reduce the
negative log-likelihood by a factor of \code{reltol * (abs(val) + reltol)}
after \code{steptol = 50} steps, with \code{reltol = 1e-9} (or if the default itermax
of 200 steps is reached).
Similarly, \link[dfoptim:nmkb]{dfoptim::nmkb}
will stop if the absolute difference of the log-likelihood between
successive iterations is below \code{tol = 1e-6}.
See \link[DEoptim:DEoptim.control]{DEoptim::DEoptim.control} and the details of \link[dfoptim:nmkb]{dfoptim::nmkb} for
further information.}
}
\description{
Find the 'best' parameter settings by fitting a \link{drift_dm} models'
predicted probability density functions to the observed data
stored within the respective object. The fitting procedure is done by
minimizing the negative log-likelihood of the model.

Users have three options:
\itemize{
\item Estimate the parameters via Differential Evolution (Default)
\item Estimate the parameters via (bounded) Nelder-Mead
\item Use Differential Evolution followed by Nelder-Mead.
}

More detailed information on how to use this function can be found in
\link{vignette("use_ddm_models", "dRiftDM")}
}
