% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_estimate_wrapper.R
\name{estimate_model_subjects}
\alias{estimate_model_subjects}
\title{Fit multiple subjects and save results}
\usage{
estimate_model_subjects(
  drift_dm_obj,
  obs_data_subject,
  lower,
  upper,
  fit_procedure_name,
  folder_name = fit_procedure_name,
  seed = NULL,
  fit_dir = "drift_dm_fits",
  force_refit = FALSE,
  progress = 2,
  ...
)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm} that will
be estimated for each subject in \code{obs_data_subject}.}

\item{obs_data_subject}{data.frame, providing an \code{RT}, \code{Error}, \code{Cond}
and \code{Subject} column. The first three columns must be suitable for
\link{set_obs_data}. The column \code{Subject} identifies a single individual}

\item{lower, upper}{numeric vectors, providing the parameter space, see
\link{estimate_model}.}

\item{fit_procedure_name}{character, providing a name of the fitting
procedure. This name will be stored in \code{drift_dm_fit_info.rds} to identify
the fitting procedure, see also \link{load_fits_subjects}.}

\item{folder_name}{character, a folder name for storing all the individual
model fits. This variable should just state the name, and should not be
a path. Per default \code{folder_name} is identical to \code{fit_procedure_name}.}

\item{seed}{numeric, a seed to make the fitting procedure reproducable
(only relevant for differntial evolution, see \link{estimate_model}).
Default is \code{NULL} which means no seed.}

\item{fit_dir}{character, a directory where (multiple) fitting procedures
can be stored. If the directory does not exist yet, it will be created
via \code{base::create.dir(fit_dir, recursive = TRUE)}. Default is
\code{"drift_dm_fits"}.}

\item{force_refit}{logical, if \code{TRUE} each subject of a fitting routine will
be fitted once more. Default is \code{FALSE} which indicates that saved files}

\item{progress}{numerical, indicating if and how progress shall be depicted.
If 0, no progress is shown. If 1, the currently fitted subject is printed
out. If 2, a progressbar is shown. Default is 2.}

\item{...}{additional arguments passed down to \link{estimate_model}.}
}
\description{
Provides a wrapper around \link{estimate_model} to fit multiple
subjects. Each subject will be stored in a folder. This folder will
also contain a file \code{drift_dm_fit_info.rds}, containing the main arguments
of the function call.
}
\details{
Examples and more information can be found here
\code{vignette("use_ddm_models", "dRiftDM")}.

When developing the fitting routine we had three levels of files/folders
in mind:
\itemize{
\item In a directory/folder named \code{fit_dir} multiple fitting routines can be
stored
\item Each fitting routine has its own folder with name as given by
\code{folder_name}
\item Within each folder, a file called
\code{drift_dm_fit_info.rds} contains the main information about the function
call. That is, the time when last modifying/calling a fitting routine, the
\code{lower} and \code{upper} parameter boundaries, the \code{drift_dm_object} that was
fitted to each individual, the original data set \code{obs_data_subject}, and
the identifier \code{fit_procedure_name}. In the same folder each individual
has its own \verb{<subject>.rds} file containing the modified \code{drift_dm_object}.
}
}
