% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_flex_prms.R
\name{flex_fix_prms}
\alias{flex_fix_prms}
\title{Exclude parameters from being modified (i.e., fix it; internal docu)}
\usage{
flex_fix_prms(flex_prms_obj, formula_instr)
}
\arguments{
\item{flex_prms_obj}{a flex_prms object}

\item{formula_instr}{a string referring to "fix"
(see \link{modify_flex_prms})}
}
\value{
a modified flex_prms_obj with respect to the (linear) internal list
}
\description{
This function modifies the (linear) internal list and sets the desired
parameters (based on the instruction string) to 0. This indicates that
this parameter is not altered within the function x2prms_vals
(i.e., ' prm <!> conda')
}
\keyword{internal}
