% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_flex_prms.R
\name{flex_restrain_prms}
\alias{flex_restrain_prms}
\title{Set parmaeters as equal across conditions (internal docu)}
\usage{
flex_restrain_prms(flex_prms_obj, formula_instr)
}
\arguments{
\item{flex_prms_obj}{flex_prms object}

\item{formula_instr}{instruction (see the notes of flex_prms)}
}
\value{
a modified flex_prms object
}
\description{
This function takes a flex_prms object and modifies the (linear) internal
list so that a parameter is set as equal across multiple conditions,
according to the instruction formula (i.e., ' prm ~! conda + condb)
}
