% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_flex_prms.R
\name{flex_special_dependency}
\alias{flex_special_dependency}
\title{Set special dependencies (internal docu)}
\usage{
flex_special_dependency(flex_prms_obj, formula_instr)
}
\arguments{
\item{flex_prms_obj}{a flex_prms object}

\item{formula_instr}{a string referring to "special dependency"
(see \link{modify_flex_prms})}
}
\value{
a modified flex_prms_object with a modified (linear) internal list
}
\description{
Sets special dependencies so that parameters depend on other parameters.
(i.e., 'prmX ~ conda == -(prmY ~ condb)')
}
