% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dm_fits_subjects.R
\name{gather_parameters}
\alias{gather_parameters}
\title{Get model values across subjects}
\usage{
gather_parameters(fits_subjects)
}
\arguments{
\item{fits_subjects}{an object inheriting from \code{dm_fits_subjects}, see
\link{load_fits_subjects}

@return
\code{gather_parameters} returns the model parameters, the log-likelihood, and
AIC/BIC values as a data.frame}
}
\description{
\verb{gather_*} functions iterate over the individual model fits in
\code{fits_subjects}, returning a data.frame of the respective parameters/values
}
