% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_bayes_estimate.R
\name{get_subset_chains}
\alias{get_subset_chains}
\title{Extract a Subset of MCMC Chains}
\usage{
get_subset_chains(chains_obj, id = NULL)
}
\arguments{
\item{chains_obj}{an object of class\code{mcmc_dm}.}

\item{id}{an optional single numeric or character, specifying the ID of a
participant to extract individual-level samples from a hierarchical model.
Ignored for non-hierarchical models.}
}
\value{
A 3D array of MCMC samples. The first dimension indicates parameters,
the second dimension chains, and the third dimension iterations
}
\description{
When calling \code{\link[=estimate_model_bayesian]{estimate_model_bayesian()}}, the MCMC results are
packed up as an \code{mcmc_dm} object. This function is used in the depths of
\code{dRiftDM} to extract the relevant array of MCMC samples,
depending on whether the model is hierarchical and whether a participant ID
is provided.
}
\keyword{internal}
