% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_solve.R
\name{im_zero}
\alias{im_zero}
\title{Derive PDFs via the Integration Method}
\usage{
im_zero(t_vec, prms_solve, one_set_comp_vecs)
}
\arguments{
\item{t_vec}{the time space}

\item{prms_solve}{the discretization, \link{prms_solve}}

\item{one_set_comp_vecs}{list of vectors containing the values for each
model component (i.e., "mu_vals", "b_vals", etc.)}
}
\value{
a list of PDFs for one condition "pdf_u" and "pdf_l"
}
\description{
Currently only a dummy function, with slow R code... Will be ported to
C++ and will be more general
}
