% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_flex_prms.R
\name{internal_list_to_matrix}
\alias{internal_list_to_matrix}
\title{Turn an internal list to a matrix}
\usage{
internal_list_to_matrix(internal_list)
}
\arguments{
\item{internal_list}{the internal list of a flex_prms object}
}
\value{
a character matrix with all parameter labels (columns) and conditions
(rows)
}
\description{
this function wrangles the internal list to character matrix
to show how each parameter relates across conditions
}
\keyword{internal}
