% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_flex_prms.R
\name{is_empty}
\alias{is_empty}
\title{Check if a vector is either empty or only contains one entry with ""
(internal docu)}
\usage{
is_empty(x)
}
\arguments{
\item{x}{a vector}
}
\value{
TRUE or FALSE otherwise
}
\description{
Check if a vector is either empty or only contains one entry with ""
(internal docu)
}
