% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_solve.R
\name{kfe_ale}
\alias{kfe_ale}
\title{Derive PDFs via the KFE}
\usage{
kfe_ale(x_vec, t_vec, prms_solve, one_set_comp_vecs)
}
\arguments{
\item{x_vec}{the evidence space}

\item{t_vec}{the time space}

\item{prms_solve}{the discretization, \link{prms_solve}}

\item{one_set_comp_vecs}{list of vectors containing the values for each
model component (i.e., "mu_vals", "b_vals", etc.)}
}
\value{
a list of PDFs for one condition "pdf_u" and "pdf_l"
}
\description{
The function calls the C++ implementation of the KFE approach by
Richter et al. and subsequently passes the first passage times forward
to \link{add_residual}.
}
