% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_flex_prms.R
\name{linearize_internal_list}
\alias{linearize_internal_list}
\title{Relabel the internal list (internal docu)}
\usage{
linearize_internal_list(internal_list)
}
\arguments{
\item{internal_list}{the internal list, with entries for each parameter x
condition combination}
}
\value{
another list, but with remapped digits, while leaving expressions or digits
of 0 untouched.
}
\description{
The entries of the internal list are either digits (0-x) or expressions.
To ensure a valid mapping of these values to an input vector (as done when an
optimizer provides input parameters), we have to linearlize the list.
}
