% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_estimate_wrapper.R
\name{load_fits_subjects}
\alias{load_fits_subjects}
\title{Load estimates of a fit procedure}
\usage{
load_fits_subjects(
  path = "drift_dm_fits",
  fit_procedure_name = "",
  detailed_info = F,
  check_data = T
)
}
\arguments{
\item{path}{character, a path pointing to a folder or directory
containing the individual model fits.}

\item{fit_procedure_name}{character, an optional name that identifies the
fit procedure that should be loaded}

\item{detailed_info}{logical, controls the amount of information displayed
in case multiple fit procedures were found and the user is prompted to
explicitly choose one}

\item{check_data}{logical, should the data be checked before passing them
back? This checks the observed data and the properties of the model. This
can take some time though. Default is \code{TRUE}}
}
\value{
An object of type \code{dm_fits_subjects} which essentially is a list with two
entries:
\itemize{
\item \code{drift_dm_fit_info}, containing a list of the main arguments when
\link{estimate_model_subjects} was originally called, including
a time-stamp
\item \code{all_fits}, containing a list of all the modified/fitted \code{drift_dm}
objects. The list's entry are named according to the subjects' identifier
}
}
\description{
This function loads the results of a fit procedure where a model was fitted
to multiple individuals
}
\details{
with respect to the logic outlined in the details of
\link{estimate_model_subjects} on the organization of fit procedures,
\code{path} could either point to a directory with (potentially) multiple fit
routines or to a specific folder with the invidiual fits. In either case
the intended location is recursively searched for files named
\code{drift_dm_fit_info.rds}.

If the fit procedure was uniquely located, either because only one fit
routine was found in the intended location or because only one
\code{drift_dm_fit_info.rds} contains the optional identifier specified in
\code{fit_procedure_name}, then all individual model fits including the
information \code{fit_procedure_name} are loaded and returned

In case multiple fit procedures are identified, the user is
prompted with a \link[utils:menu]{utils::menu}, listing information about the possible
candidates. The intended fit procedure can then interactively be chosen
by the user. The amount of displayed information is controlled via
\code{detailed_info}.
}
