% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_s3_methods.R
\name{logLik.drift_dm}
\alias{logLik.drift_dm}
\title{Extract Log-Likelihood for a drift_dm Object}
\usage{
\method{logLik}{drift_dm}(object, ...)
}
\arguments{
\item{object}{a \link{drift_dm} object containing observed data}

\item{...}{additional arguments}
}
\value{
A \code{logLik} object containing the log-likelihood value for the
\link{drift_dm} object. This value has attributes for the number of
observations (\code{nobs}) and the number of model parameters (\code{df}).

Returns \code{NULL} if the observed data is not available.
}
\description{
This method extracts the log-likelihood for a \code{drift_dm} object, ensuring
data is available and evaluating the model if necessary.
}
