% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_flex_prms.R
\name{max_number_one_internal_entry}
\alias{max_number_one_internal_entry}
\title{Get the maximum number from an internal entry (internal docu)}
\usage{
max_number_one_internal_entry(one_internal_entry)
}
\arguments{
\item{one_internal_entry}{one entry of multiple conditions}
}
\value{
the largest digit in this entry (0 if there are only expressions)
}
\description{
The entries of the internal list are either digits (0-x) or expressions.
}
