% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_flex_prms.R
\name{modify_flex_prms}
\alias{modify_flex_prms}
\title{Modify the flex_prms object using the user instructions (internal docu)}
\usage{
modify_flex_prms(flex_prms_obj, instr, messaging = T)
}
\arguments{
\item{flex_prms_obj}{an object of type flex_prms}

\item{instr}{a character string with the instructions; with "\n" separating
each instruction}

\item{messaging}{shall additional messages be displayed?}
}
\value{
an updated flex_prms object
}
\description{
Modify the flex_prms object using the user instructions (internal docu)
}
