% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dm.R
\name{mu.drift_dm}
\alias{mu.drift_dm}
\alias{mu_int.drift_dm}
\alias{x.drift_dm}
\alias{b.drift_dm}
\alias{dt_b.drift_dm}
\alias{nt.drift_dm}
\title{Fallback functions for the diffusion model components}
\usage{
\method{mu}{drift_dm}(drift_dm_obj, t_vec, one_cond)

\method{mu_int}{drift_dm}(drift_dm_obj, t_vec, one_cond)

\method{x}{drift_dm}(drift_dm_obj, x_vec, one_cond)

\method{b}{drift_dm}(drift_dm_obj, t_vec, one_cond)

\method{dt_b}{drift_dm}(drift_dm_obj, t_vec, one_cond)

\method{nt}{drift_dm}(drift_dm_obj, t_vec, one_cond)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}}

\item{t_vec}{numeric vector defining the time space (from 0 to \code{t_max} in
steps of \code{dt}; see \link{drift_dm})}

\item{one_cond}{a character of length 1 indicating a condition (see
\link{drift_dm})}

\item{x_vec}{numeric vector defining the standardized evidence space (from
-1 to 1 in steps of \code{dx}; see \link{drift_dm})}
}
\value{
\itemize{
\item \code{mu.drift_dm} (drift rate) returns a vector of length \code{length(t_vec)}
containing the number \eqn{3}.
\item \code{mu_int.drift_dm} (integral of the drift rate) returns a vector of length
\code{length(t_vec)} with the values \code{t_vec*3}.
\item \code{x.drift_dm} (starting point distribution) returns a dirac delta on zero
with length \code{length(x_vec)}. Note that if \code{nx} (see \link{drift_dm}) is
even, the dirac delta will be slightly shifted to a value > 0.
\item \code{b.drift_dm} (boundary) returns a vector of length \code{length(t_vec)} with
the values \eqn{0.5}.
\item \code{dt_b.drift_dm} (derivative of the boundary) returns a vector of length
\code{length(t_vec)} with the values \eqn{0}.
\item \code{nt.drift_dm} (non-decision time) returns a dirac delta on \eqn{t = 0.3}
with length \code{length(t_vec)}. Note that if \eqn{t = 0.3} is not provided in
\code{t_vec}, the dirac delta is centered on the closest time step below
\eqn{t = 0.3}.
}
}
\description{
the provided methods with the ending \code{.drift_dm} will be called in two cases.
\itemize{
\item First, if users explicitly pass a \code{drift_dm_obj} of type
\link{drift_dm} to  \link{mu}, \link{mu_int}, \link{x},
\link{b}, \link{dt_b}, or \link{nt}. For instance, when
creating a bare bone \link{drift_dm} object and then deriving model
predictions.
\item Second, in case a user defines an object inheriting from \code{drift_dm},
without specifying custom versions of the respective methods. See
\link{vignette("use_ddm_models", "dRiftDM")} for more information.
}

In other words: The functions with the ending \code{.drift_dm} are fall-back
methods which provide fixed settings for the different diffusion model
components in case no class-specific methods can be found.
}
