% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dm.R
\name{mu}
\alias{mu}
\alias{mu_int}
\alias{x}
\alias{b}
\alias{dt_b}
\alias{nt}
\title{Functions for defining the components of a diffusion model}
\usage{
mu(drift_dm_obj, t_vec, one_cond)

mu_int(drift_dm_obj, t_vec, one_cond)

x(drift_dm_obj, x_vec, one_cond)

b(drift_dm_obj, t_vec, one_cond)

dt_b(drift_dm_obj, t_vec, one_cond)

nt(drift_dm_obj, t_vec, one_cond)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}}

\item{t_vec}{numeric vector defining the time space (from 0 to \code{t_max} in
steps of \code{dt}; see \link{drift_dm})}

\item{one_cond}{a character of length 1 indicating a condition (see
\link{drift_dm})}

\item{x_vec}{numeric vector defining the standardized evidence space (from
-1 to 1 in steps of \code{dx}; see \link{drift_dm})}
}
\value{
The values returned by the functions depend on the class of
their first argument. See the documentation of the particular methods for
details of what is returned by that method.

In order to work with \code{dRiftDM}, \code{mu}, \code{mu_int}, \code{b}, \code{dt_b}, \code{nt} must
return a numeric vector of the same length as \code{t_vec}, providing the
respective values at every time step \eqn{t}.

In order to work with \code{dRiftDM}, \code{x} must return a numeric vector of the
same length as \code{x_vec}.

See \link{vignette("use_ddm_models", "dRiftDM)} for more information how each
function works and how to write a customized method for your own models
}
\description{
\code{mu} and \code{mu_int} provide the drift rate and its integral, respectively,
across the time space in a condition.

\code{x} provides a distribution of the starting point across the evidence
space.

\code{b} and \code{dt_b}provide the values of the (symmetric) boundary and its
derivative, respectively, across the time space in a condition.

\code{nt} provides a distribution of the non-decision component across the
time space in one condition.

All of the listed functions are generic functions which means that
custom methods can be defined for new \link{class}es.
}
\details{
Please visit the \link{vignette("use_ddm_models", "dRiftDM)} for more in-depth
information.
\subsection{Drift rate and its integral:}{

The drift rate is the first derivative of the expected time-course
of the diffusion process. For instance, if we assume that the diffusion
process \eqn{X} is linear with a slope of \eqn{v}...
\deqn{E(X) = v \cdot t}
...then the drift rate at every time step \eqn{t} is the constant \eqn{v},
obtained by taking the derivative of the expected time-course with respect
to \eqn{t}:
\deqn{\mu(t) = v}
Conversely, the integral of the drift rate is identical to the expected
time-course:
\deqn{\mu_{int}(t) = v \cdot t}

For the drift rate \code{mu}, the default method for a model of type
\link{drift_dm} will return a numeric vector of length \code{length(t_vec)}
containing the number \eqn{3}. Its integral counterpart \code{mu_int} will return
a numeric vector containing the values \code{t_vec*3} (see
\link{mu.drift_dm}).
}

\subsection{Starting Point Distribution:}{

The starting point of a diffusion model refers to the initial value taken
by the evidence accumulation process at time \eqn{t=0}.

The default method for a model of type \link{drift_dm} returns a dirac
delta on zero, meaning that every potential diffusion process starts at 0
(see \link{x.drift_dm}).
}

\subsection{Boundary:}{

The Boundary refers to the values of the absorbing boundaries at every time
step \eqn{t} in a diffusion model. In most cases, this will be a constant.
For instance:
\deqn{b(t) = b}
In this case, its derivative with respect to \eqn{t} is 0.

The default method for a model of type \link{drift_dm} will return a
numeric vector of length \code{length(t_vec)} containing the number \eqn{0.5}.
Its counterpart \code{dt_b} will return a numeric vector of the same length
containing its derivative, namely, \code{0} (see \link{b.drift_dm}).
}

\subsection{Non-Decision Time:}{

The non-decision time refers to an additional time-requirement. Its
distribution across the time space will be convoluted with the pdfs derived
from the diffusion process.

In psychology, the non-decision time captures time-requirements outside the
central decision process, such as stimulus perception and motor execution.

The default method for a model of type \link{drift_dm} returns a dirac
delta, shifted to \eqn{t = 0.3} (see \link{nt.drift_dm}).
}
}
