% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{mu_constant}
\alias{mu_constant}
\alias{mu_int_constant}
\alias{b_constant}
\alias{nt_constant}
\title{Basic drift diffusion model components}
\usage{
mu_constant(drift_dm_obj, t_vec, one_cond)

mu_int_constant(drift_dm_obj, t_vec, one_cond)

b_constant(drift_dm_obj, t_vec, one_cond)

nt_constant(drift_dm_obj, t_vec, one_cond)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}.}

\item{t_vec}{numeric vector providing the time space}

\item{one_cond}{character, indicating a condition}
}
\description{
The following functions can be used to modify/customize the components of
a diffusion model

\code{mu_constant}, this function provides the component function for a constant
drift rate with parameter \code{muc}.

\code{mu_int_constant}, provides the complementary integral to \code{mu_constant}.

\code{b_constant}, this function provides the component function for a constant
boundary with parameter \code{b}.

\code{nt_constant}, this function provides the component function for a constant
non-decision time with parameter \code{non_dec}.
}
\details{
See \code{vignette("use_ddm_models", "dRiftDM")} for more information on how
to set/modify/customize the components of a diffusion model.
}
