% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{mu_dmc}
\alias{mu_dmc}
\title{Drift Rate for DMC}
\usage{
mu_dmc(prms_model, prms_solve, t_vec, one_cond, ddm_opts)
}
\arguments{
\item{prms_model}{the model parameters, containing muc, tau, a, A}

\item{prms_solve}{solver settings}

\item{t_vec}{time space}

\item{one_cond}{one condition}

\item{ddm_opts}{optional arguments attached to an object}
}
\value{
provides the first derivative of the superimposed process with
respect to t_vec.
}
\description{
Provides the drift rate of the superimposed decision process. That is
the derivative of the rescaled gamma function plus a constant drift rate for
the controlled process.
}
\keyword{internal}
