% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{mu_dmc_dm}
\alias{mu_dmc_dm}
\alias{mu_int_dmc_dm}
\title{Drift Rate for DMC}
\usage{
mu_dmc_dm(drift_dm_obj, t_vec, one_cond)

mu_int_dmc_dm(drift_dm_obj, t_vec, one_cond)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}.}

\item{t_vec}{numeric vector providing the time space}

\item{one_cond}{character, indicating a condition}
}
\description{
The following functions are used for the drift rate of DMC
(see \link{dmc_dm}).

\code{mu_dmc_dm}, this function provides the superimposed diffusion process
of DMC. Necessary parameters \code{muc}, \code{a}, \code{A}, \code{tau}.

\code{mu_int_dmc_dm}, provides the complementary integral to \code{mu_dmc_dm}.
}
\details{
See \code{vignette("use_ddm_models", "dRiftDM")} for more information on how
to set/modify/customize the components of a diffusion model.
}
