% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{mu_int_dmc}
\alias{mu_int_dmc}
\title{Integral of DMC's Drift Rate}
\usage{
mu_int_dmc(prms_model, prms_solve, t_vec, one_cond, ddm_opts)
}
\arguments{
\item{prms_model}{the model parameters, containing muc, tau, a, A}

\item{prms_solve}{solver settings}

\item{t_vec}{time space}

\item{one_cond}{one condition}

\item{ddm_opts}{optional arguments attached to an object}
}
\value{
provides the scaled gamma distribution function of the superimposed
process for each time step in t_vec.
}
\description{
Provides the integral of the drift rate of the superimposed decision process.
This is the sum of the rescaled gamma function and the linear function.
}
\keyword{internal}
