% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{mu_ssp}
\alias{mu_ssp}
\title{Drift Rate for SSP}
\usage{
mu_ssp(prms_model, prms_solve, t_vec, one_cond, ddm_opts)
}
\arguments{
\item{prms_model}{the model parameters, containing p, sd_0, r, sign}

\item{prms_solve}{solver settings}

\item{t_vec}{time space}

\item{one_cond}{one condition}

\item{ddm_opts}{optional arguments attached to an object}
}
\value{
provides the drift rate for SSP with respect to t_vec
}
\description{
Provides the drift rate for the SSP model. That is, the sum of attention
attributed to the flankers and central target, scaled by the perceptual
input.
}
