% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{nt_constant}
\alias{nt_constant}
\title{Constant Non-Decision time}
\usage{
nt_constant(prms_model, prms_solve, t_vec, one_cond, ddm_opts)
}
\arguments{
\item{prms_model}{the model parameters; containing "non_dec"}

\item{prms_solve}{solver settings}

\item{t_vec}{time space}

\item{one_cond}{one condition}

\item{ddm_opts}{optional arguments attached to an object}
}
\value{
a vector of the same length as t_vec with zeros, except for the
element matching with "non_dec" with respect to "t_vec"
}
\description{
A dirac delta on "non_dec", to provide a constant non-decision time.
}
