% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{nt_truncated_normal}
\alias{nt_truncated_normal}
\title{Truncated Normally-Distributed Non-Decision Time}
\usage{
nt_truncated_normal(prms_model, prms_solve, t_vec, one_cond, ddm_opts)
}
\arguments{
\item{prms_model}{the model parameters; including "non_dec" and
"sd_non_dec"}

\item{prms_solve}{solver settings}

\item{t_vec}{time space}

\item{one_cond}{one condition}

\item{ddm_opts}{optional arguments attached to an object}
}
\value{
returns the PDF of a truncated normal distribution for t_vec, with
mean "non_dec" and standard deviation "sd_non_dec". Lower truncation is 0.
Upper truncation is max(t_vec)
}
\description{
Truncated Normally-Distributed Non-Decision Time
}
\keyword{internal}
