% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{nt_uniform}
\alias{nt_uniform}
\title{Uniform Non-Decision Time}
\usage{
nt_uniform(prms_model, prms_solve, t_vec, one_cond, ddm_opts)
}
\arguments{
\item{prms_model}{the model parameters; including "non_dec" and
"range_non_dec"}

\item{prms_solve}{solver settings}

\item{t_vec}{time space}

\item{one_cond}{one condition}

\item{ddm_opts}{optional arguments attached to an object}
}
\value{
returns the PDF of a uniform distribution for t_vec, centered around
"non_dec" and with a range of "range_non_dec".
}
\description{
Uniform Non-Decision Time
}
