% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_model_comps}
\alias{plot_model_comps}
\title{Visualize Model Components}
\usage{
plot_model_comps(drift_dm_obj, line_cols = NULL, xlim_time = NULL)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}}

\item{line_cols}{character vector, providing colors for each condition in
\code{drift_dm_obj$conds}. Default is \code{NULL}, meaning that colors are determined by
\link[grDevices:palettes]{grDevices::rainbow}}

\item{xlim_time}{numeric vector of length 2, providing the limits of the
x-axis of all plots related to the time space. Default is \code{NULL} which
refers to \eqn{[0, t_{max}/2]}}
}
\description{
This function takes an object of type \link{drift_dm} and plots the
output of each component function in \code{drift_dm_obj$comp_funs}
}
