% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_quantiles}
\alias{plot_quantiles}
\title{Plot the quantiles}
\usage{
plot_quantiles(
  drift_dm_obj,
  type = "both",
  probs = seq(0.1, 0.9, 0.1),
  to_plot = "Quant_Corr",
  x_lab = NULL,
  y_lab = NULL,
  x_lim = NULL,
  y_lim = NULL,
  line_cols = NULL
)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}}

\item{type}{character, indicating whether CAFs of the model ("pred"), the
observed data ("obs"), or both ("both") should be plotted. Default is
"both".}

\item{probs}{numeric vector, providing the quantiles to plot. Default is
\code{seq(0.1, 0.9, 0.1)}}

\item{to_plot}{character, providing the column of the data.frame returned by
\link{calc_quantiles} that should be used for plotting. Default is
"Quant_Corr", indicating that the quantiles of correct responses will be
plotted.}

\item{x_lab, y_lab}{character, providing a label for the x-axis and y-axis,
respectively. Default is "RT" and "F(RT)", respectively.}

\item{x_lim, y_lim}{numeric vectors of length 2, providing the limits of the
x-axis and y-axis. The default for \code{x_lim} is \code{c(0, t_max / 2)}, with
\code{t_max} defined within \code{drift_dm_obj}. The default for \code{y_lim} is
\code{c(0, 1)}.}

\item{line_cols}{character vector, indicating the color-coding of the
conditions. Default colors are based on the \code{grDevices::rainbow} palette.}
}
\description{
This function provides a basic plot of the model's and/or the observed data's
quantiles (see \link{calc_quantiles}).
}
