% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot trace(s) of a model}
\usage{
plot_trace(
  drift_dm_obj,
  add_x = TRUE,
  sigma = NULL,
  k = 1,
  x_lab = NULL,
  y_lab = NULL,
  x_lim = NULL,
  y_lim = NULL,
  line_cols_ev = NULL,
  line_cols_b = NULL,
  seed = NULL
)
}
\arguments{
\item{drift_dm_obj}{an object of type drift_dm}

\item{add_x}{logical, indicating whether a starting point should be
added. Default is TRUE.}

\item{sigma}{numerical, for controlling the diffusion parameter when
simulating traces. Set 0 to plot the expected time-course of the diffusion
model. The default (NULL) leaves the diffusion constant of \code{drift_dm_obj}
unchanged.}

\item{k}{numerical, number of traces to plot. Default is 1.}

\item{x_lab, y_lab}{character, providing a label for the x-axis and y-axis,
respectively. Default is "Time" and "Evidence", respectively.}

\item{x_lim, y_lim}{numeric vectors of length 2, providing the limits of the
x-axis and y-axis. The default for \code{x_lim} is \code{c(0, t_max / 4)}, with
\code{t_max} defined within \code{drift_dm_obj}. The default for \code{y_lim} is the
maximum of the boundary function provided in \code{drift_dm_obj}.}

\item{line_cols_ev}{character vector, indicating the color-coding of the
conditions. Default colors are based on the \code{grDevices::rainbow} palette.}

\item{line_cols_b}{character, providing a color for the boundary. Default
is black.}

\item{seed}{a seed for making the simulated traces reproducable. Default
(NULL) uses no seed.}
}
\description{
This functions creates a basic plot that depicts a user-defined number
of traces (see \link{simulate_trace}).
May come in handy when developing/testing/exploring a model.
}
