% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_traces}
\alias{plot_traces}
\title{Plot trace(s) of a model}
\usage{
plot_traces(
  drift_dm_obj,
  k,
  conds = NULL,
  add_x = FALSE,
  seed = NULL,
  sigma = NULL,
  x_lab = NULL,
  y_lab = NULL,
  x_lim = NULL,
  y_lim = NULL,
  line_cols_ev = NULL,
  line_cols_b = NULL
)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}}

\item{k}{numeric, the number of traces to simulate per condition}

\item{conds}{character vector, conditions for which traces shall be simulated}

\item{add_x}{logical, indicating whether a starting point should be
added. Default is FALSE}

\item{seed}{numerical, an optional seed for reproducible sampling}

\item{sigma}{numerical, for controlling the diffusion parameter when
simulating traces. The default (NULL) leaves the diffusion constant of
\code{drift_dm_obj} unchanged.}

\item{x_lab, y_lab}{character, providing a label for the x-axis and y-axis,
respectively. Default is "Time" and "Evidence", respectively.}

\item{x_lim, y_lim}{numeric vectors of length 2, providing the limits of the
x-axis and y-axis. The default for \code{x_lim} is \code{c(0, t_max / 4)}, with
\code{t_max} defined within \code{drift_dm_obj}. The default for \code{y_lim} is the
maximum of the boundary function provided in \code{drift_dm_obj}.}

\item{line_cols_ev}{character vector, indicating the color-coding of the
conditions. Default colors are based on the \code{grDevices::rainbow} palette.}

\item{line_cols_b}{character, providing a color for the boundary. Default
is black.}
}
\description{
This functions creates a basic plot that shows a user-defined number
of traces (see \link{simulate_trace}).
May come in handy when developing/testing/exploring a model.
}
