% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_stats_dm.R
\name{print.cafs}
\alias{print.cafs}
\alias{print.quantiles}
\alias{print.delta_funs}
\alias{print.fit_stats}
\alias{print.sum_dist}
\title{Print Functions for Stats Objects}
\usage{
\method{print}{cafs}(x, ...)

\method{print}{quantiles}(x, ...)

\method{print}{delta_funs}(x, ...)

\method{print}{fit_stats}(x, ...)

\method{print}{sum_dist}(x, ...)
}
\arguments{
\item{x}{a subclass of \link{data.frame}, as returned by \code{\link[=calc_stats]{calc_stats()}}.}

\item{...}{additional arguments passed forward to
\code{\link[=print.stats_dm]{print.stats_dm()}}.}
}
\value{
\code{x} (invisibly)
}
\description{
when calling \code{\link[=calc_stats]{calc_stats()}}, each returned statistic will be a
subclass of \code{stats_dm} and \link{data.frame}. The following \code{print() }methods will
call the more generic \code{\link[=print.stats_dm]{print.stats_dm()}} function.
}
\keyword{internal}
