% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dm.R
\name{prms_cond_combo}
\alias{prms_cond_combo}
\title{Unique Conditions-Parameter Combinations}
\usage{
prms_cond_combo(drift_dm_obj)
}
\arguments{
\item{drift_dm_obj}{an object of type \link{drift_dm}}
}
\value{
a matrix with two rows. Each column contains a combination of the
parameter name and the condition that can be considered unique. Parameter
names are stored in the first row, condition labels in the second.
}
\description{
This is a helper function. It searches through the \code{linear_internal_list} of
the stored \link{flex_prms} object, and keeps the first unique appearance
of parameters. For example, when the parameter muc is equal for comp, neutral,
and incomp, the function will provide the info "muc" and "comp", thus
dropping incomp and neutral, where the parameter is the same.
}
