% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_flex_prms.R
\name{prms_conds_to_modify}
\alias{prms_conds_to_modify}
\title{Extract the conditions and parameters from an instruction string}
\usage{
prms_conds_to_modify(formula_instr, operation, all_conds, all_prms)
}
\arguments{
\item{formula_instr}{an instruction string}

\item{operation}{what to expect in terms of the string's structure. Can be
"vary", "restrain", "fix", "set" or "dependency"}

\item{all_conds}{all potential conditions (necessary for extending missing
condition specification)}

\item{all_prms}{all potential paramters (necessary for extending missing
condition specification)}
}
\value{
a named list prms_to_adress and conds_to_adress as character
vectors.
}
\description{
This function takes an instruction as a string and then extracts the
conditions and parameters, depending on what "type" of operation it is.
The expected structure is listed in the details docu of flex_prms()
}
