% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ratcliff_dm}
\alias{ratcliff_dm}
\title{Create a basic Diffusion Model}
\usage{
ratcliff_dm(
  var_non_dec = FALSE,
  var_start = FALSE,
  obs_data = NULL,
  sigma = 1,
  t_max = 3,
  dt = 0.001,
  dx = 0.001
)
}
\arguments{
\item{var_non_dec, var_start}{logical, indicating whether the model should have
a (uniform) variable non-decision time or starting point, respectively
(see \code{nt_uniform} and \code{x_uniform} in \link{component_shelf})}

\item{obs_data}{data.frame, an optional data.frame with the observed data.
See \link{set_obs_data}.}

\item{sigma, t_max, dt, dx}{numeric, providing the settings for the diffusion
constant and discretization (see \link{drift_dm})}
}
\description{
This function creates a \link{drift_dm} that corresponds to the basic
Ratcliff diffusion model
}
\details{
The Ratcliff diffusion model is a diffusion model with  a constant drift rate
\code{muc} and a constant boundary \code{b}. If \code{var_non_dec = FALSE},  a constant
non-decision time \code{non_dec} is assumed, otherwise a uniform non-decision time
with mean \code{non_dec} and range \code{range_non_dec}. If \code{var_start = FALSE},  a
constant starting point centered between the boundaries is assumed (i.e.,
a dirac delta over 0), otherwise a uniform starting point with mean 0 and
range \code{range_start}.
}
