% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ratcliff_synth_data}
\alias{ratcliff_synth_data}
\title{A synthetic data set with one condition}
\format{
A data frame with 300 rows and 3 columns:
\describe{
\item{RT}{Response Times}
\item{Error}{Error Coding (Error Response = 1; Correct Response = 0)}
\item{Cond}{Condition ('null')}
}
}
\usage{
ratcliff_synth_data
}
\description{
This dataset was simulated by using the classical ratcliff diffusion model
(see  \code{\link[=ratcliff_dm]{ratcliff_dm()}}).
}
\keyword{datasets}
