% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dm.R
\name{remove_estimate_info}
\alias{remove_estimate_info}
\title{Remove flags added when calling estimate_classical}
\usage{
remove_estimate_info(drift_dm_obj)
}
\arguments{
\item{drift_dm_obj}{a \link{drift_dm} object}
}
\value{
the model without the list entry \code{estimate_info}
}
\description{
This is a small internal function that ensures that the flags added
after estimating a model via \code{\link[=estimate_classical]{estimate_classical()}} are removed. It
is used when replacing settings of a model after it has been estimated.
}
\keyword{internal}
