% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dm.R
\name{set_one_solver_setting}
\alias{set_one_solver_setting}
\title{Set one specific aspect of the solver settings}
\usage{
set_one_solver_setting(drift_dm_obj, name_prm_solve, value_prm_solve)
}
\arguments{
\item{drift_dm_obj}{an object of type \link{drift_dm}}

\item{name_prm_solve}{which aspect to address? ("sigma", "t_max", "dx", "dt",
"solver")}

\item{value_prm_solve}{either a single numeric or character string}
}
\value{
the updated un-evauated (!) drift_dm_obj object
}
\description{
Internal function to update one aspect of \code{prms_solve} or \code{solver}.
}
\details{
Ensures that the supplied values are reasonable and that \code{nx} and \code{nt} are
updated. The functions \link{prms_solve<-} and
\link{solver<-} pass their arguments forward to this function.
}
