% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{set_plot_limits}
\alias{set_plot_limits}
\title{Set Plot Limits}
\usage{
set_plot_limits(lim, default_lim)
}
\arguments{
\item{lim}{numeric vector of length 2, specifying the desired plot limits.}

\item{default_lim}{numeric vector of length 2, default limits to use if
\code{lim} is NULL.}
}
\value{
A numeric vector of length 2, specifying the plot limits.
}
\description{
This function determines plot limits, using \code{default_lim} if \code{lim} is not
specified.
}
