% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simon_data}
\alias{simon_data}
\title{A simulated data set with two conditions}
\format{
\subsection{\code{ratcliff_data}}{

A data frame with 300 rows and 3 columns:
\describe{
\item{RT}{Response Times}
\item{Error}{Error Coding (Error Response = 1; Correct Response = 0)}
\item{Cond}{Condition ('comp' and 'incomp')}
}
}
}
\usage{
simon_data
}
\description{
This dataset was simulated by using the diffusion model for conflict tasks
(see  \code{\link[=dmc_dm]{dmc_dm()}}) with parameter settings that are typical for a
Simon task.
It will be used for creating examples in the dRiftDM package...
}
\keyword{datasets}
