% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dm.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate "observed" response times from a diffusion model}
\usage{
simulate_data(drift_dm_obj, n, seed = NULL)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}.}

\item{n}{numeric, the number of samples per condition to draw}

\item{seed}{numeric, an optional seed for reproducable sampling}
}
\value{
a data.frame containing the columns \code{RT}, \code{Error}, and \code{Cond}
}
\description{
Cdfs are derived from the model's pdfs and response times are drawn by
mapping samples from a uniform distribution (in \eqn{[0, 1]}) to the values
of the cdf. Note that sampled response times will correspond to the
the values of the time space (i.e., they will correspond to
\code{seq(0, t_max, dt)}, see \link{drift_dm}).
}
\details{
This function simulates data based on the provided diffusion model. To this
end, random samples from the model's predicted pdfs are drawn via approximate
inverse cdf sampling.
}
