% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dm.R
\name{simulate_trace}
\alias{simulate_trace}
\title{Simulate trajectories/traces of a model}
\usage{
simulate_trace(drift_dm_obj, k, one_cond, add_x = FALSE, seed = NULL)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}.}

\item{k}{numeric, the number of traces to simulate}

\item{one_cond}{character, a condition for which traces shall be simulated}

\item{add_x}{logical, indicating whether traces should contain a variable
starting point. If \code{TRUE}, samples from \code{b_fun} (see \link{drift_dm}) are
drawn and added to each trace. Default is \code{FALSE}.}

\item{seed}{numerical, an optional seed for reproducable sampling}
}
\value{
For \code{k = 1}, the function returns a single vector of length \code{nt + 1} (where
\code{nt} is the number of steps in the discretization of time; see
\link{drift_dm}).

For \code{k = l} (with \code{l > 1}), the function returns a matrix of size
\verb{(l, nt + 1)}.

Evidence values where traces went beyond the boundary of the
model are set to NA before passing them back.
}
\description{
This function simulates single trajectories/traces of a diffusion model
(i.e., single evidence accumulation processes) using forward euler.

Might come in handy when exploring the model's behavior or when
creating figures (see also \link{plot_trace})
}
\details{
The algorithm for simulating traces is forward euler. See
\insertCite{Richteretal.2023;textual}{dRiftDM} and
\insertCite{Ulrichetal.2015;textual}{dRiftDM} (Appendix A) for more
information
}
\references{
\insertAllCited{}
}
