% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dm.R
\name{simulate_traces}
\alias{simulate_traces}
\title{Simulate trajectories/traces of a model}
\usage{
simulate_traces(drift_dm_obj, k, conds = NULL, add_x = FALSE, seed = NULL)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}.}

\item{k}{numeric, the number of traces to simulate per condition}

\item{conds}{character vector, conditions for which traces shall be
simulated}

\item{add_x}{logical, indicating whether traces should contain a variable
starting point. If \code{TRUE}, samples from \code{x_fun} (see \link{drift_dm}) are
drawn and added to each trace. Default is \code{FALSE}.}

\item{seed}{numerical, an optional seed for reproducible sampling}
}
\value{
The structre of the returned object depends on the arguments provided.
If conds is a character vector of length 1 (i.e., a single string), either
a single vector of length \code{nt + 1} or a matrix of size (\code{k}, \code{nt + 1})
is returned (depending on whether \code{k} = 1 or \code{k} > 1, respectively; note
that \code{nt} is the number of steps in the discretization of time;
see \link{drift_dm}).

If conds is a character vector of length > 1, a named list is returned with
either vectors or matrices as entries.

Note that evidence values with traces beyond the boundary of the
model are set to NA before passing them back.
}
\description{
This function simulates single trajectories/traces of a diffusion model
(i.e., single evidence accumulation processes) using forward euler.

Might come in handy when exploring the model's behavior or when
creating figures (see also \link{plot_traces})
}
\details{
The algorithm for simulating traces is forward euler. See
\insertCite{Richteretal.2023;textual}{dRiftDM} and
\insertCite{Ulrichetal.2015;textual}{dRiftDM} (Appendix A) for more
information
}
