% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_flex_prms.R
\name{sort_one_internal_entry}
\alias{sort_one_internal_entry}
\title{Sorts the numbers in ascending order (internal docu)}
\usage{
sort_one_internal_entry(one_internal_entry)
}
\arguments{
\item{one_internal_entry}{one entry of a (linearlized) internal_list}
}
\value{
the newly sorted entry as a list
}
\description{
.... within one entry (i.e., one prm
across all conditions, such as when calling internal_list[\link{"A"}]); which
is a list. Used to ensure that parameter ordering remains logical
}
\details{
The entries of the internal list are either digits (0-x) or expressions.
}
