% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ssp_dm}
\alias{ssp_dm}
\title{Create the Shrinking Spotlight Model}
\usage{
ssp_dm(obs_data = NULL, sigma = 1, t_max = 3, dt = 0.001, dx = 0.001)
}
\arguments{
\item{obs_data}{data.frame, an optional data.frame with the observed data.
See \link{set_obs_data}.}

\item{sigma, t_max, dt, dx}{numeric, providing the settings for the diffusion
constant and discretization (see \link{drift_dm})}
}
\description{
This function creates a \link{drift_dm} object that corresponds to a
simple version of the shrinking spotlight model by
\insertCite{Whiteetal.2011;textual}{dRiftDM}.
}
\details{
The shrinking spotlight model is a model developed for the flanker task.

It has the following properties (see \link{component_shelf}):
\itemize{
\item a constant boundary (parameter \code{b})
\item a constant starting point in between the decision boundaries
\item an evidence accumulation process that is driven by an attentional
spotlight that covers both the flankers and the target. The area that covers
the flankers and target is modeled by normal distribution with mean 0:
\itemize{
\item At the beginning of the trial attention is wide-spread, and the width
at t=0 is the standard deviation \code{sd_0}
\item As the trial progresses in time, the attentional spotlight narrows,
reflected by a linear decline of the standard deviation with rate \code{r}
(to a minimum of 0.001).
\item the attention attributed to both the flankers and the target is scaled
by \code{p} which controls the strength of evidence accumulation
}
\item A non-decision time that follows a truncated normal distribution with
mean \code{non_dec} and standard deviation \code{sd_non_dec}.
}

Its parameters are:  \code{b}, \code{non_dec}, \code{sd_non_dec}, \code{p}, \code{sd_0}, and \code{r}.
Per default, all parameters but \code{r} are assumed to be "free".
}
\references{
\insertRef{Whiteetal.2011}{dRiftDM}
}
