% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ssp_synth_data}
\alias{ssp_synth_data}
\title{A synthetic data set with two conditions}
\format{
A data frame with 600 rows and 3 columns:
\describe{
\item{RT}{Response Times}
\item{Error}{Error Coding (Error Response = 1; Correct Response = 0)}
\item{Cond}{Condition ('comp' and 'incomp')}
}
}
\usage{
ssp_synth_data
}
\description{
This dataset was simulated by using the Shrinking Spotlight Model
(see \code{\link[=ssp_dm]{ssp_dm()}}) with parameter settings that are typical for a
Flanker task.
}
\keyword{datasets}
