% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_estimate2.R
\name{summarize_drift_dm_info}
\alias{summarize_drift_dm_info}
\title{Reads Info file}
\usage{
summarize_drift_dm_info(full_name_to_file, detailed_info)
}
\arguments{
\item{full_name_to_file}{path to the info file (.RDS)}

\item{detailed_info}{logical, if detailed info shall be provided or not}
}
\value{
a string with infos about the fit procedure name, last call,
and, if detailed_info = TRUE, model, individuals, lower/upper, and seed.
}
\description{
Requires a path to the info file of a fit procedure (see
\link{estimate_model_ids}) and turns its main information into a string
}
\keyword{internal}
