% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_s3_methods.R
\name{try_cast_integer}
\alias{try_cast_integer}
\title{Convert Character Digits to Numeric Digits}
\usage{
try_cast_integer(values)
}
\arguments{
\item{values}{a vector of values to attempt conversion to integer.}
}
\value{
an integer vector if conversion succeeds; otherwise, the original
vector.
}
\description{
This internal function casts a character vector to integer, if the character
vector only contains digits for entries. If the input vector is not of
type character or if any entry contains a non-digit, then the vector is
returned unmodified.
}
\keyword{internal}
