% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_s3_methods.R
\name{try_cast_numeric}
\alias{try_cast_numeric}
\title{Convert Digit to Digits Numeric}
\usage{
try_cast_numeric(values)
}
\arguments{
\item{values}{a vector of values to attempt conversion to numeric.}
}
\value{
a numeric vector if conversion succeeds; otherwise, the original
vector is returned.
}
\description{
This internal function casts a character vector to numeric, if the character
vector only contains digits for entries. If the input vector is not of
type character or if any entry contains a non-digit, then the vector is
returned unmodified.
}
