% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_flex_prms.R
\name{update_special_values}
\alias{update_special_values}
\title{Update all prms}
\usage{
update_special_values(flex_prms_obj)
}
\arguments{
\item{flex_prms_obj}{a flex_prms object}
}
\value{
the modified flex_prms_obj (i.e,. with the updated prms_matrix and
the updated cust_prms$values)
}
\description{
This function takes a flex_prms object and updates the prms_matrix according
to the special instructions in \code{internal_list} and the custom parameters
\code{cust_prms}
}
\keyword{internal}
