% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_estimate2.R
\name{validate_fits_ids}
\alias{validate_fits_ids}
\title{Validate a an Object of Type fits_ids_dm}
\usage{
validate_fits_ids(fits_ids, progress)
}
\arguments{
\item{fits_ids}{an object of type fits_ids_dm}

\item{progress}{numeric, 0, no progress, 1 basic output, 2 progress bars}
}
\value{
the unmodified fits_ids objects after passing all tests.
}
\description{
checks if all the information are in the fits_ids_dm (see
\link{estimate_model_ids}) object and ensures that nothing obviously
fishy is going on with the individual model fits.
}
\details{
Checks:
\itemize{
\item if all names are in the info file
\item and if the respective entries make sense
\item if the flex_prms object of the all saved models and the overall
model is the same (except for the differences in the prm values).
\item if the class, prms_solve and solver of the saved models and the
overall model is the same
\item if the estimated parameters are in the parameter space
\item for same  b_coding and same functions
\item if the number of fits matches with the number of individuals
in the info file
\item if the data in each fitted model matches with the observed data in
the info file
}
}
\keyword{internal}
