% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_flex_prms.R
\name{validate_flex_prms}
\alias{validate_flex_prms}
\title{checks if a flex_prms_object is parameterized in a reasonable way}
\usage{
validate_flex_prms(flex_prms_obj)
}
\arguments{
\item{flex_prms_obj}{and object of type flex_prms}
}
\value{
the unmodified flex_prms_obj object; only stop/warning statements
}
\description{
\itemize{
\item checks internal_list
\item checks liner_internal_list
\item checks the parameter matrix
}
}
