% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_flex_prms.R
\name{x2prms_vals}
\alias{x2prms_vals}
\title{Update the parameter matrix for vector inputs (internal docu)}
\usage{
x2prms_vals(x, flex_prms_obj)
}
\arguments{
\item{x}{a numeric vector with new values to set}

\item{flex_prms_obj}{a flex_prms_obj with the (linearized) internal list and
the parameter matrix}
}
\value{
a flex_prms_obj with updated parameter matrix
}
\description{
This function takes a numeric vector and maps the values to the parameter
matrix using the linearlized internal list. This will also lead to an update
of the values for which special dependencies were set
}
