% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{x_beta}
\alias{x_beta}
\title{A beta-shaped starting point component}
\usage{
x_beta(drift_dm_obj, x_vec, one_cond)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}.}

\item{x_vec}{numeric vector providing the evidence space}

\item{one_cond}{character, indicating a condition}
}
\description{
This function provides the function component for a symmetric beta-shaped
starting point distribution with parameter \code{alpha}.
}
\details{
See \code{vignette("use_ddm_models", "dRiftDM")} for more information on how
to set/modify/customize the components of a diffusion model.
}
