% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{x_dirac_0}
\alias{x_dirac_0}
\title{Constant Starting Point at Zero}
\usage{
x_dirac_0(prms_model, prms_solve, x_vec, one_cond, ddm_opts)
}
\arguments{
\item{prms_model}{the model parameters; no prm name required}

\item{prms_solve}{solver settings}

\item{x_vec}{evidence space}

\item{one_cond}{one condition}

\item{ddm_opts}{optional arguments attached to an object}
}
\value{
a vector of the same length as x_vec with zeros, except for the
element in the middle of the vector
}
\description{
A dirac delta on zero, to provide no bias and a constant starting point
}
