% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{x_uniform}
\alias{x_uniform}
\title{Uniform Starting Point Distribution Centered Around Zero}
\usage{
x_uniform(prms_model, prms_solve, x_vec, one_cond, ddm_opts)
}
\arguments{
\item{prms_model}{the model parameters; prm name "range_start" required}

\item{prms_solve}{solver settings}

\item{x_vec}{evidence space}

\item{one_cond}{one condition}

\item{ddm_opts}{optional arguments attached to an object}
}
\value{
returns the PDF of a uniform distribution for x_vec, centered around
zero and with a range of "range_start".
}
\description{
Uniform Starting Point Distribution Centered Around Zero
}
