% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_solve.R
\name{calc_pdfs}
\alias{calc_pdfs}
\title{Calculate the PDFs}
\usage{
calc_pdfs(solver, x_vec, t_vec, prms_solve, one_set_comp_vecs)
}
\arguments{
\item{solver}{a single character, specifying the approach to derive
the first passage time (i.e., "kfe" or "im_zero")}

\item{x_vec}{numeric vector, the evidence space}

\item{t_vec}{numeric vector, the time space}

\item{prms_solve}{the discretization (see \link{prms_solve})}

\item{one_set_comp_vecs}{a list with all vectors for each model component
(i.e., mu_vals, b_vals etc.)}
}
\value{
a list of PDFs for one condition "pdf_u" and "pdf_l"
}
\description{
This function takes the necessary vectors to derive the first passage
time (for one condition; subsetted before calling this function). It only
calls further functions, and doesn't perform any calculations
}
\seealso{
\code{\link[=kfe_ale]{kfe_ale()}}, \code{\link[=im_zero]{im_zero()}}, \link{add_residual}
}
