% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_solve.R
\name{calc_pdfs}
\alias{calc_pdfs}
\title{Derive the PDFs of a model}
\usage{
calc_pdfs(drift_dm_obj, one_cond, solver)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}}

\item{one_cond}{character, providing the name of a condition}

\item{solver}{character, indicating which algorithm should be used for
deriving the pdf. Default is to "kfe", meaning that the Crank-Nicolson
discretization of the Kolmogorov-Forward-Equation should be used.}
}
\description{
This function calculates the probability density functions of a drift
diffusion model for one condition.
}
\details{
More in-depth information about the mathematical details can be found in
\insertCite{Richteretal.2023;textual}{dRiftDM}
}
\references{
\insertAllCited{}
}
