% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_stats.R
\name{calc_stats}
\alias{calc_stats}
\title{Calcuating Statistics}
\usage{
calc_stats(drift_dm_obj, type, source = "both", ...)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}}

\item{type}{character vector, indicating which statistics should be
calculated. Currently supported options are "cafs", "quantiles", "delta_funs".}

\item{source}{character, indicating whether statistics for observed data ("obs"),
the model's predictions ("pred"), or both ("both") should be calculated.
Default is "both".}

\item{...}{additional optional and necessary parameters passed down to the
functions handling the calculations. See Details for more information}
}
\value{
The return value depends on the \code{type} argument. If \code{type} is  a character
vector of length 1, a single data.frame is returned containing the requested
statistics. If \code{type} is a character vector of length > 1, a named list
is passed back, with each entry in the list corresponding to a data.frame of
statistics.
}
\description{
This function is a wrapper for calculating statistics on model predictions
or observed data. Currently supported are:
\itemize{
\item Conditional Accuracy Functions (CAFs)
\item Quantiles
\item Delta Functions
}
}
\section{Conditional Accuracy Function (CAFs)}{
CAFs are a way to quantify response accuracy against speed. To calculate
CAFs, RTs (whether correct or incorrect) are first binned and then the
percent correct responses per bin is calculated.

When calculating model-based CAFs, a joint cdf combining both the pdf
of correct and incorrect responses is calculated. Afterwards, this cdf
is separated into even-spaced segments and the contribution of
the pdf associated with a correct response relative to the joint cdf is
calculated.

The number of bins can be controlled by passing the argument \code{n_bins}.
The default is 5.
}

\section{Quantiles}{
For observed response times, the function \link[stats:quantile]{stats::quantile} is used with
default settings.

Which quantiles are calcuated can be controlled by providing the
probabilites, \code{probs}, with values in \eqn{[0, 1]}. Default is
\code{seq(0.1, 0.9, 0.1)}.
}

\section{Delta Functions}{
Delta functions calculate the difference between quantiles
of two conditions against their mean:
\itemize{
\item \eqn{Delta_i = Q_{i,j} - Q_{i,k}}
\item \eqn{Avg_i = 0.5 \cdot Q_{i,j} + 0.5 \cdot Q_{i,k}}
}

With i indicating a quantile, and j and k two conditions.

To calculate delta functions, users have to specify:
\itemize{
\item \code{minuends}: character vector, specifying condition(s) j. Must be in
\code{drift_dm_obj$conds}.
\item \code{subtrahends}: character vector, specifying condition(s) k. Must be in
\code{drift_dm_obj$conds}
\item \code{dvs}: character vector, specifying the dependent variable(s) to use.
Supported are \code{Quant_Corr} and \code{Quant_Err}
\item optional: \code{probs}, see the section on quantiles
}
}

