% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_flex_prms.R
\name{check_internal_list}
\alias{check_internal_list}
\title{checks if all entries of internal_list are an expression or integer.
Throws an error if not (internal docu)}
\usage{
check_internal_list(internal_list, prm_names, cond_names)
}
\arguments{
\item{internal_list}{a list, referring to a (linear) internal list of
a \link{flex_prms} object.}

\item{prm_names}{the expected parameter of the flex_prms object}

\item{cond_names}{the expected conditions of the flex_prms object}
}
\value{
the internal list for convenience
}
\description{
Checks also if there are entries for each parameter and condition
}
