% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dm.R
\name{comp_vals}
\alias{comp_vals}
\title{Evaluate all Component Functions}
\usage{
comp_vals(
  drift_dm_obj,
  x_vec = NULL,
  t_vec = NULL,
  nx = NULL,
  nt = NULL,
  dx = NULL,
  dt = NULL,
  t_max = NULL
)
}
\arguments{
\item{drift_dm_obj}{an object of type \link{drift_dm}}

\item{x_vec}{optional, the discretized evidence space}

\item{t_vec}{optional, the discretized time space}

\item{nx, nt, dx, dt}{optional, the steps and step sizes of each space}

\item{t_max}{the maximum time of the time space.}
}
\value{
If solver "kfe", a named list with entries "mu_vals", "x_vals", "b_vals",
"dt_b_vals", "nt_vals".

If solver "im_*", the returned list will also contain "mu_int_vals".
}
\description{
Gets/calculates all values provided by the component functions of a
\link{drift_dm} object
}
\details{
arguments are optional, because they can be extracted from the model.
However, supplying these are faster than creating them.
}
