% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dm.R
\name{drift_dm}
\alias{drift_dm}
\title{Create a drift_dm object}
\usage{
drift_dm(
  prms_model,
  conds,
  free_prms = NULL,
  obs_data = NULL,
  sigma = 1,
  t_max = 3,
  dt = 0.001,
  dx = 0.001
)
}
\arguments{
\item{prms_model}{A named numeric vector of the model parameters. The names
indicate the model's parameters, and the numeric entries provide the current
parameter values.}

\item{conds}{A character vector, giving the names of the model's conditions.
values within \code{conds} will be used when addressing the data and when
deriving the model's predictions}

\item{free_prms}{A character vector indicating which of the parameters
defined in \code{prms_model} are allowed to vary (i.e., can be
estimated/changed). The entries of \code{free_prms} have to match
\code{names(prms_model)}. The default (\code{NULL}) will set
\code{free_prms <- names(prms_model)}}

\item{obs_data}{A \code{data.frame} used to fill the \code{drift_dm} object with data.
See \link{set_obs_data} for more information.}

\item{sigma}{The diffusion constant. Default is set to 1.}

\item{t_max}{The maximum of the time space. Default is set to 3 (seconds).}

\item{dt}{The step size of the time discretization. Default is set to .001
(seconds).}

\item{dx}{The step size of the evidence space discretization. Default is set
to .001.}
}
\value{
A list with the class label "drift_dm". In general, it is not
recommended to directly modify the entries of this list. Users should use the
built-in setter functions (e.g., \link{set_model_prms});
see \code{vignette("use_ddm_models", package = "dRiftDM")} for more
information). The list contains the following entries:
\itemize{
\item The named numeric vector \code{prms_model}
\item The character vector \code{conds}
\item Parameters used for deriving the model predictions, \code{prms_solve},
containing the diffusion constant (\code{sigma}), the maximum of the time space
(\code{t_max}), the steps of time and evidence space (\code{dt} and \code{dx},
respectively), and the number of steps in the time and evidence space
(\code{nt} and \code{nx}, respectively).
\item The character vector \code{free_prms}
\item A label indicating the method for deriving the model predictions
(currently only \code{kfe}).
}

If observed data were passed to the model, either when calling \code{drift_dm()}
or when calling \link{set_obs_data}, the list will contain an entry
called \code{obs_data}. \code{obs_data}, in turn, is again a list, containing the
response times of the correct or erroneous responses across conditions.

If the model has been evaluated (see \link{re_evaluate_model}), the
list will additionally contain...
\itemize{
\item ... the log likelihood;can be addressed via \code{drift_dm_obj$log_like_val}
\item ... the AIC/BIC values; can be addressed via \code{drift_dm_obj$ic_vals}
}
}
\description{
This function creates an object of type drift_dm, which serves as the parent
class for all further created drift diffusion models. Its structure is the
backbone  of the dRiftDM package and every child of the drift_dm class must
have the attributes  of the parent class. Typically, users will not want to
create an object of drift_dm alone, as its use is very limited. Rather, they
will want an object of one of its  child classes. See
\link{vignette("use_ddm_models", "dRiftDM)} for a list of pre-built diffusion
models and more information on how to create and use child classes.
}
