% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dm.R, R/formatting_drift_dm.R
\name{drift_dm}
\alias{drift_dm}
\alias{print.drift_dm}
\title{Create a drift_dm object}
\usage{
drift_dm(
  prms_model,
  conds,
  subclass,
  instr = NULL,
  obs_data = NULL,
  sigma = 1,
  t_max = 3,
  dt = 0.001,
  dx = 0.001,
  solver = "kfe",
  mu_fun = NULL,
  mu_int_fun = NULL,
  x_fun = NULL,
  b_fun = NULL,
  dt_b_fun = NULL,
  nt_fun = NULL,
  b_coding = NULL
)

\method{print}{drift_dm}(x, ..., round_digits = drift_dm_default_rounding())
}
\arguments{
\item{prms_model}{a named numeric vector of the model parameters. The names
indicate the model's parameters, and the numeric entries provide the current
parameter values.}

\item{conds}{a character vector, giving the names of the model's conditions.
values within \code{conds} will be used when addressing the data and when
deriving the model's predictions.}

\item{subclass}{a character string, with a name for the newly created
diffusion model (e.g., \code{dmc_dm}). This will be the child class.}

\item{instr}{an optional character string, providing "instructions" for the
underlying \link{flex_prms} object.}

\item{obs_data}{an optional data.frame, providing a data set (see
\code{\link[=obs_data]{obs_data()}} for more information).}

\item{sigma}{the diffusion constant. Default is \code{1}.}

\item{t_max}{the maximum of the time space. Default is set \code{3} (seconds).}

\item{dt, dx}{the step size of the time and evidence space discretization,
respectively. Default is set to \code{.001} (which refers to seconds for dt).
Note that these values are set conservatively per default. In many cases,
users can increase the discretization.}

\item{solver}{a character string, specifying which approach to use for
deriving the first passage time. Default is \code{kfe}, which provides access to
the numerical discretization of the Kolmogorov Forward Equation.}

\item{mu_fun, mu_int_fun, x_fun, b_fun, dt_b_fun, nt_fun}{Optional custom
functions defining the components of a diffusion model. See
\code{\link[=comp_funs]{comp_funs()}}.}

\item{b_coding}{an optional list, specifying how boundaries are coded. See
\code{\link[=b_coding]{b_coding()}}. Default refers to accuracy coding.}

\item{x}{an object of type \code{drift_dm}}

\item{...}{additional parameters}

\item{round_digits}{integer, controls the number of digits shown for
\code{\link[=print.drift_dm]{print.drift_dm()}}. Default is \code{3}.}
}
\value{
A list with the parent class label \code{"drift_dm"} and the child class
label \verb{<subclass>}. The list contains the following entries:
\itemize{
\item An instance of the class \link{flex_prms} for controlling the model
parameters. Provides information about the number of parameters, conditions
etc.
\item Parameters used for deriving the model predictions, \link{prms_solve},
containing the diffusion constant (\code{sigma}), the maximum of the time space
(\code{t_max}), the evidence and space discretization (\code{dt} and \code{dx},
respectively), and the resulting number of steps for the time and evidence
space discretization (\code{nt} and \code{nx}, respectively).
\item A character string \code{solver}, indicating the method for deriving the model
predictions.
\item A list of functions called \link{comp_funs}, providing the components
of the diffusion model (i.e., \code{mu_fun}, \code{mu_int_fun}, \code{x_fun}, \code{b_fun},
\code{dt_b_fun}, \code{nt_fun}). These functions are called in the depths of the
package and will determine the behavior of the model
}

If (optional) observed data were passed via \code{\link[=obs_data]{obs_data()}},
the list will contain an entry \code{obs_data}. This is a (nested) list with
stored response times for the upper and lower boundary and with respect to
each condition.

If the model has been evaluated (see \code{\link[=re_evaluate_model]{re_evaluate_model()}}), the
list will additionally contain...
\itemize{
\item ... the log likelihood; can be addressed via \code{\link[=logLik.drift_dm]{logLik.drift_dm()}}.
\item ... the PDFs of the first passage time; can be addressed via
\code{drift_dm_obj$pdfs}.
}

Every model also has the attribute \link{b_coding}, which summarizes how
the boundaries are labeled.
}
\description{
This function creates an object of type drift_dm, which serves as the parent
class for all further created drift diffusion models. Its structure is the
backbone of the dRiftDM package and every child of the drift_dm class must
have the attributes of the parent class. Typically, users will not want to
create an object of drift_dm alone, as its use is very limited. Rather, they
will want an object of one of its child classes. See
\code{vignette("create_ddm_models", "dRiftDM")} for more information on how
to create/use/modify child classes.
}
\details{
In general, it is not recommended to directly modify the
entries of this list unless you are familiar with how the package works.
Users should use the replacement methods and the
\code{\link[=modify_flex_prms]{modify_flex_prms()}} method to modify the model. See
\code{vignette("use_ddm_models", "dRiftDM")} and
\code{vignette("create_ddm_models", "dRiftDM")} for more information.
}
\seealso{
\code{\link[=conds]{conds()}}, \code{\link[=flex_prms]{flex_prms()}}, \code{\link[=prms_solve]{prms_solve()}},
\code{\link[=solver]{solver()}}, \code{\link[=obs_data]{obs_data()}}, \code{\link[=comp_funs]{comp_funs()}},
\code{\link[=b_coding]{b_coding()}}, \code{\link[=coef]{coef()}}
}
