% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRiftDM-package.R
\docType{package}
\name{dRiftDM-package}
\alias{dRiftDM}
\alias{dRiftDM-package}
\title{dRiftDM: Fitting and Exploring Drift Diffusion Models}
\description{
This package will help users to fit a drift diffusion model with a time-dependent drift rate. The goal is to provides useful functions to easily implement, fit, and explore a drift diffusion model. The package also comes with pre-implemented drift diffusion models for fitting data from conflict tasks.
}
\author{
\strong{Maintainer}: Valentin Koob \email{koob@uni-bremen.de}

Authors:
\itemize{
  \item Thomas Richter
  \item Markus Janczyk
}

}
\keyword{internal}
