% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRiftDM-package.R
\docType{package}
\name{dRiftDM-package}
\alias{dRiftDM}
\alias{dRiftDM-package}
\title{dRiftDM: Fitting and Exploring Drift Diffusion Models}
\description{
Drift Diffusion models (DDMs) are common in cognitive psychology to desribe decision processes in simple tasks. We provide a wrapper around the numerical solution by Richter et al. (2023) to easy apply DDMs via Maximum-Likelihood-Estimation.
}
\author{
\strong{Maintainer}: Valentin Koob \email{koob@uni-bremen.de}

Authors:
\itemize{
  \item Thomas Richter
  \item Markus Janczyk
}

}
\keyword{internal}
