% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_starting_values}
\alias{get_starting_values}
\alias{get_starting_values.drift_dm}
\title{TITLE}
\usage{
get_starting_values(object, ...)

\method{get_starting_values}{drift_dm}(
  object,
  ...,
  lower = NULL,
  upper = NULL,
  verbose = 0,
  use_ez = NULL,
  n_lhs = NULL
)
}
\arguments{
\item{object}{}

\item{...}{}
}
\description{
TITLE
}
\keyword{internal}
