% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_numeric}
\alias{is_numeric}
\title{Check if an object is a valid numeric vector}
\usage{
is_numeric(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
A logical value: \code{TRUE} if the input is a numeric vector without any
missing or infinite values, otherwise \code{FALSE}.
}
\description{
This function verifies whether the input is a numeric vector with no missing
(\code{NA}, \code{NaN}) or infinite (\code{Inf} or \code{-Inf}) values.
}
\keyword{internal}
