% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_flex_prms.R
\name{max_number_one_internal_entry}
\alias{max_number_one_internal_entry}
\alias{get_number_prms}
\title{Get the maximum number from an internal entry or flex_prms_obj
(internal docu)}
\usage{
max_number_one_internal_entry(one_internal_entry)

get_number_prms(flex_prms_obj)
}
\arguments{
\item{one_internal_entry}{one entry of multiple conditions}

\item{flex_prms_obj}{a list stored in a flex_prms object}
}
\value{
the largest digit in the entry or of the linear_list in
the supplied flex_prms_obj (0 if there are only
expressions). The largest number of the linear_list corresponds to the number
of model parameters.
}
\description{
The entries of the internal list are either digits (0-x) or expressions.
}
