% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{mu_constant}
\alias{mu_constant}
\title{Constant Drift Rate}
\usage{
mu_constant(prms_model, prms_solve, t_vec, one_cond, ddm_opts)
}
\arguments{
\item{prms_model}{the model parameters, containing muc}

\item{prms_solve}{solver settings}

\item{t_vec}{time space}

\item{one_cond}{one condition}

\item{ddm_opts}{optional arguments attached to an object}
}
\value{
a vector of the same length as t_vec with the drift rate for
each element of the vector.
}
\description{
Constant Drift Rate
}
\keyword{internal}
