% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{nt_truncated_normal}
\alias{nt_truncated_normal}
\title{A truncated-normal shaped non-decision time component}
\usage{
nt_truncated_normal(drift_dm_obj, t_vec, one_cond)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}.}

\item{t_vec}{numeric vector providing the time space}

\item{one_cond}{character, indicating a condition}
}
\description{
This function provides the function component for a non-decision time
following a truncated-normal distribution with mean \code{non_dec} and
standard deviation \code{sd_non_dec}.
}
\details{
See \code{vignette("use_ddm_models", "dRiftDM")} for more information on how
to set/modify/customize the components of a diffusion model.
}
