% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cafs}
\alias{plot_cafs}
\title{Plot conditional accuracy functions}
\usage{
plot_cafs(
  drift_dm_obj,
  type = "both",
  n_bins = 5,
  x_lab = NULL,
  y_lab = NULL,
  x_lim = NULL,
  y_lim = NULL,
  line_cols = NULL
)
}
\arguments{
\item{drift_dm_obj}{an object of type drift_dm}

\item{type}{character, indicating whether CAFs of the model ("pred"), the
observed data ("obs"), or both ("both") should be plotted. Default is
"both".}

\item{n_bins}{numeric, indicating how many bins the CAF should contain.
Default is 5.}

\item{x_lab, y_lab}{character, providing a label for the x-axis and y-axis,
respectively. Default is "Bins" and "Accuracy \link[=\\\%]{\\%}", respectively.}

\item{x_lim, y_lim}{numeric vectors of length 2, providing the limits of the
x-axis and y-axis. The default for \code{x_lim} is \code{c(1, n_bins)}. The default
for \code{y_lim} is \code{c(0, 1)}.}

\item{line_cols}{character vector, indicating the color-coding of the
conditions. Default colors are based on the \code{grDevices::rainbow} palette.}
}
\description{
This function provides a basic plot of the model's and/or the observed data's
conditional accuracy functions (CAFs; see \link{calc_cafs}).
}
