% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_prms}
\alias{plot_prms}
\title{Plot Parameter Distribution(s)}
\usage{
plot_prms(fits_subjects, include_fit_values = F, col = "skyblue")
}
\arguments{
\item{fits_subjects}{an object inheriting from \code{dm_fits_subjects} (see
\link{load_fits_subjects})}

\item{include_fit_values}{logical, indicating if a histogram for
\code{log_like}, \code{AIC}, and \code{BIC} should be created as well}

\item{col}{the color of the histrogram bars}
}
\description{
This function creates a histogram for each parameter in \code{fits_subjects}
}
\details{
This function uses \link{gather_parameters} to gather parameters across
subjects for a single object of type \code{dm_fits_subjects}. Subsequently it
creates a histogram for each parameter. The final plot has multiple panels,
each for one parameter.
}
