% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ratcliff_dm}
\alias{ratcliff_dm}
\title{Create a basic Diffusion Model}
\usage{
ratcliff_dm(
  type = "simple",
  obs_data = NULL,
  sigma = 1,
  t_max = 3,
  dt = 0.001,
  dx = 0.001
)
}
\arguments{
\item{type}{character, indicatnig "which" Ratcliff diffusion model should
be build. Default and only option currently is "simple".}

\item{obs_data}{data.frame, an optional data.frame with the observed data.
See \link{set_obs_data}.}

\item{sigma, t_max, dt, dx}{numeric, providing the settings for the
discretization (see \link{drift_dm})}
}
\description{
This function creates a \link{drift_dm} that corresponds to the basic
Ratcliff diffusion model
}
\details{
\subsection{Simple Model}{

With "simple" Ratcliff Diffusion Model we refer to a diffusion model
with  a constant drift rate \code{muc}, a constant boundary \code{b}, a constant
non-decision time \code{non_dec}, and no starting point variability. See also
\link{mu_constant}.
}
}
