% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dm.R
\name{re_evaluate_model}
\alias{re_evaluate_model}
\title{Re-evaluate the model}
\usage{
re_evaluate_model(drift_dm_obj)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}}
}
\value{
Returns the passed \code{drift_dm_obj} object, after (re-)calculating
the log likelihood (and AIC/BIC) of the model.
\itemize{
\item the log likelihood can be addressed via \code{drift_dm_obj$log_like_val}
\item the AIC/BIC values can be addressed via \code{drift_dm_obj$ic_vals}
}
}
\description{
Updates the log likelihood value stored in \code{drift_dm_obj} and subsequently
calculates the values for the Akaike information criterion (AIC) and the
Bayesian information criterion (BIC)
}
