% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{set_default_arguments}
\alias{set_default_arguments}
\title{Set default graphical parameters for plotting in dRiftDM}
\usage{
set_default_arguments(dots, conds)
}
\arguments{
\item{dots}{a named list of graphical parameters (typically created from
\code{...} when passing additional arguments to \verb{plot.*()} functions in \code{dRiftDM}).}
}
\value{
A list with all updated graphical parameters
}
\description{
This function sets graphical parameters (passed through \code{...} by
\code{\link[=plot.cafs]{plot.cafs()}}, \code{\link[=plot.quantiles]{plot.quantiles()}},
\code{\link[=plot.delta_funs]{plot.delta_funs()}}, and \code{\link[=plot.densities]{plot.densities()}}) to sensible
defaults used in plotting. It supports arguments relevant to
\code{\link[graphics:plot.default]{graphics::plot()}}, \code{\link[graphics:points]{graphics::points()}}, and \code{\link[graphics:legend]{graphics::legend()}},
allowing users fine-grained control over plotting while avoiding argument
clashes.
}
\details{
\itemize{
\item \code{pch}: plotting symbol of points (default: \code{19})
\item \code{lty}: line type of lines (default: \code{1})
\item \code{pt.bg}: point background color (default: \code{NULL})
\item \code{lwd}: line width of lines (default: \code{1})
\item \code{cex.axis}, \code{cex.lab}, \code{cex.main}, \code{cex.pt}, \code{cex.legend}: scaling
factors for text, axes, and points (default: \code{1})
\item \code{main}: plot title (default: empty string, so no title)
\item \code{family}: font family (default: empty string)
\item \code{legend.pos}: legend position (default: \code{"bottomright"}, could either
be a string with a position as handled by the \code{x} argument of
\code{\link[graphics:legend]{graphics::legend()}} or a numeric vector of the legend's position)
\item \code{legend}: legend labels (default: falls back to the specified \code{conds})
\item \code{bg}: background color of the legend box (default: \code{"white"})
\item \code{bty}: box type for the legend (default: \code{"o"})
\item \code{border}: border color of the legend box (default: \code{"black"})
\item \code{box.lwd}: line width of the legend box (default: \code{1})
\item \code{box.col}: border color of the legend box (default: \code{"black"})
\item \code{box.lty}: line type of the legend box (default: \code{1})
\item \code{err.width}: the width of the caps at the end of each error bar
\item \code{err.col}: the color of the error bar (default: \code{NULL}, this will
set the error bar color to either the point/line color of the plot or to
gray50; depending on the type of plot)
\item \code{shade.alpha}: transparency of the shaded area around the model prediction
(default: \code{0.2})
}
}
\note{
The value of \code{legend} defaults to \code{conds}, which is internally
supplied by dRiftDM. TODO: Update
}
\keyword{internal}
