% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{set_default_colors}
\alias{set_default_colors}
\title{Set Default Colors}
\usage{
set_default_colors(colors, n)
}
\arguments{
\item{colors}{character vector, specifying colors for conditions. If \code{NULL},
default colors are used.}

\item{n}{}
}
\value{
A character vector of colors of length \code{n}.
}
\description{
This function assigns default colors if none are provided or adjusts the
color vector to match the number \code{n}
}
\keyword{internal}
