% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dm.R
\name{set_model_prms}
\alias{set_model_prms}
\alias{set_free_prms}
\alias{set_solver_settings}
\alias{set_obs_data}
\title{Setting attributes of a drift_dm model}
\usage{
set_model_prms(drift_dm_obj, new_model_prms, eval_model = F)

set_free_prms(drift_dm_obj, new_free_prms)

set_solver_settings(
  drift_dm_obj,
  names_prm_solve,
  values_prm_solve,
  eval_model = F
)

set_obs_data(drift_dm_obj, obs_data, eval_model = F)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}}

\item{new_model_prms}{a numeric vector of the same length as
\code{drift_dm_obj$free_prms}, specifying the new values of the parameters that
are allowed to vary}

\item{eval_model}{logical, indicating whether \link{re_evaluate_model}
should be called after modifying the model. Default is \code{FALSE}. Note that if
\code{eval_model} is set to \code{FALSE}, the attributes \code{log_like_val} and \code{ic_vals}
are deleted from the model. Also, \code{eval_model = TRUE} only has an effect
if the model provides data.}

\item{new_free_prms}{a character vector specifying the names of the
parameters that are allowed to vary}

\item{names_prm_solve}{a character vector, possible entries are
\code{solver}, \code{sigma}, \code{t_max}, \code{dt}, \code{dx} (Note that \code{solver} can only be
\code{kfe} at the moment).}

\item{values_prm_solve}{numeric or character, defining the values to be set
for \code{names_prm_solve}.}

\item{obs_data}{a \link{data.frame} which provides three columns: (1) \code{RT} for
the response times, (2) \code{Error} for error coding (1 = error, 0 = correct),
(3) \code{Cond} for specifying the conditions (see
\link{vignette("use_ddm_models", "dRiftDM")} for more information).}
}
\value{
each setter method passes back the modified model.
}
\description{
Functions starting with \verb{set_*} provide ways for modifying the
list underlying every object inheriting from \code{drift_dm}. Using the setter
methods is the highly recommended way of changing the attributes of a model
(see \link{drift_dm} for a list of the built-in attributes).
\itemize{
\item \code{set_model_prms} for setting the parameters of the model that are
declared as "free".
\item \code{set_free_prms} for declaring which parameters are "free" (i.e.,
are allowed to vary or be modified)
\item \code{set_solver_settings} for modifying the settings relevant to the functions
deriving the pdfs of the diffusion model (only one parameter at a time can
be changed)
\item \code{set_obs_data} can be used to pass/set observed data
}
}
