% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dm.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate "observed" response times from a diffusion model}
\usage{
simulate_data(drift_dm_obj, n, df_prms = NULL, seed = NULL, verbose = 1)
}
\arguments{
\item{drift_dm_obj}{an object inheriting from \link{drift_dm}.}

\item{n}{numeric, the number of samples per condition to draw}

\item{df_prms}{data.frame, an optional data.frame providing the parameters
that should be used for simulating the data. \code{df_prms} must provide columns
named like the parameters in \code{drift_dm_obj$free_prms}, plus a column \code{Subject}
that will identify each simulated data set.}

\item{seed}{numeric, an optional seed for reproducable sampling}

\item{verbose}{integer, indicating if information about the progress
should be displayed. 0 -> no information,
1 -> a progress bar. Default is 1. Only effective when \code{df_prms} is provided}
}
\value{
a data.frame containing at least the columns \code{RT}, \code{Error}, and \code{Cond}. If
\code{df_prms} is provided, then the data.frame will additionally contain the
column \code{Subject}
}
\description{
Cdfs are derived from the model's pdfs and response times are drawn by
mapping samples from a uniform distribution (in \eqn{[0, 1]}) to the values
of the cdf. Note that sampled response times will correspond to the
the values of the time space (i.e., they will correspond to
\code{seq(0, t_max, dt)}, see \link{drift_dm}).
}
\details{
This function simulates data based on the provided diffusion model. To this
end, random samples from the model's predicted pdfs are drawn via approximate
inverse cdf sampling.
}
