% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dm.R
\name{simulate_one_data_set}
\alias{simulate_one_data_set}
\title{Simulate one data set}
\usage{
simulate_one_data_set(drift_dm_obj, n)
}
\arguments{
\item{drift_dm_obj}{a \link{drift_dm} object}

\item{n}{numeric, specifying the number of trials per condition. Can be
single numeric, or a (named) numeric vector with the same length as
conds(drift_dm_obj)}
}
\value{
A data.frame with the columns "RT", "<b_column>", and "Cond"; and
with n rows.
}
\description{
Function that simulates a single data based on a model (using the prms set
}
\keyword{internal}
