% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_tools.R
\name{simulate_values}
\alias{simulate_values}
\title{Simulate Values}
\usage{
simulate_values(lower, upper, n, distr = "unif", seed = NULL, ...)
}
\arguments{
\item{lower, upper}{numeric vectors, indicating the lower/upper boundary of
the drawn values.}

\item{n}{numeric, the number of values to be drawn for each value pair of
lower/upper}

\item{distr}{character, indicating which distribution to draw from. Currently
available are: \code{"unif"} for a uniform distribution or \code{"tnorm"} for a
truncated normal distribution}

\item{seed}{numeric, optional seed for making the simulation reproducable}

\item{...}{further arguments relevant for the distribution to draw from
(see details)}
}
\value{
A matrix with \code{n} rows and \code{length(lower)/length(upper)} columns.
}
\description{
Draw values, most likely model parameters.
}
\details{
When drawing from a truncated normal distribution, users must provide values
for the arguments \code{means} and \code{sds}. These are numeric vectors of the same
size as \code{lower} and \code{upper}, and indicate the mean and the standard deviation
of the normal distribution.
}
