% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_s3_methods.R
\name{try_cast_numeric}
\alias{try_cast_numeric}
\title{Attempt to Convert Values to Numeric}
\usage{
try_cast_numeric(values)
}
\arguments{
\item{values}{a vector of values to attempt conversion to numeric.}
}
\value{
a numeric vector if conversion succeeds; otherwise, the original
vector is returned.
}
\description{
This internal function tries to cast a vector to numeric. If the conversion
fails (e.g., due to non-numeric values), it returns the original input
unchanged.
}
