% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ulrich_flanker_data}
\alias{ulrich_flanker_data}
\title{Exemplary Flanker Data}
\format{
A data.frame with 16 individuals and the following columns:
\describe{
\item{ID}{Individual IDs}
\item{RT}{Response Times}
\item{Error}{Error Coding (Error Response = 1; Correct Response = 0)}
\item{Cond}{Condition ('comp' and 'incomp')}
}
}
\usage{
ulrich_flanker_data
}
\description{
Data of the Flanker task collected in the course of the study by
\insertCite{Richteretal.2023;textual}{dRiftDM}.
}
\keyword{datasets}
