% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{x_beta}
\alias{x_beta}
\title{Beta-Shaped Starting Point Distribution Centered Around Zero}
\usage{
x_beta(prms_model, prms_solve, x_vec, one_cond, ddm_opts)
}
\arguments{
\item{prms_model}{the model parameters; prm name "alpha" required}

\item{prms_solve}{solver settings}

\item{x_vec}{evidence space}

\item{one_cond}{one condition}

\item{ddm_opts}{optional arguments attached to an object}
}
\value{
returns the PDF of a beta-shaped distribution for x_vec, centered
around zero and with a shape parameter "alpha".
}
\description{
Beta-Shaped Starting Point Distribution Centered Around Zero
}
